[DEV-3687] updating tests
