+++ tests/containers/award/AwardContainer-test.jsx
@@ -1,99 +1,137 @@
 /**
- * AwardContainer-test.js
- * Created by Emily Gullo 03/03/2017
- **/
+ * awardContainer-test.js
+ * Created by David Trinh 10/6/2018
+ * */
 
 import React from 'react';
-import { mount, shallow } from 'enzyme';
+import { shallow } from 'enzyme';
 
 import { AwardContainer } from 'containers/award/AwardContainer';
+import BaseContract from 'models/v2/awardsV2/BaseContract';
+import BaseIdv from 'models/v2/awardsV2/BaseIdv';
+import BaseFinancialAssistance from "models/v2/awardsV2/BaseFinancialAssistance";
 
-import { mockParams, mockActions, mockApi, mockFinancialAssistanceApi } from './mockResults';
-
-import BaseContract from 'models/v2/awards/BaseContract';
-import BaseFinancialAssistance from "models/v2/awards/BaseFinancialAssistance";
+import { mockParams, mockActions } from './mockAward';
+import { mockContract, mockLoan, mockIdv } from '../../models/awardsV2/mockAwardApi';
 
 jest.mock('helpers/searchHelper', () => require('./awardHelper'));
+jest.mock("helpers/downloadHelper", () => require("./awardHelper"));
 
-// mock the child component by replacing it with a function that returns a null element
-jest.mock('components/award/Award', () => jest.fn(() => null));
+// mock the child components by replacing them with functions that return null elements
+jest.mock('components/award/award', () => jest.fn(() => null));
+jest.mock('containers/award/AwardContainer', () => jest.fn(() => null));
 
-describe('AwardContainer', () => {
+const getAwardContainer = (params = mockParams) => shallow(<AwardContainer
+    {...params}
+    {...mockActions} />);
+
+describe('awardContainer', () => {
     it('should make an API call for the selected award on mount', async () => {
-
-        const container = mount(<AwardContainer
-            {...mockParams}
-            {...mockActions} />);
+        const container = getAwardContainer();
 
         const parseAward = jest.fn();
         container.instance().parseAward = parseAward;
+        container.instance().componentDidMount();
         await container.instance().awardRequest.promise;
 
         expect(parseAward).toHaveBeenCalled();
     });
 
     it('should make an API call when the award ID parameter changes', () => {
-        const container = shallow(<AwardContainer
-            {...mockParams}
-            {...mockActions} />);
+        const container = getAwardContainer();
 
         const getSelectedAward = jest.fn();
         container.instance().getSelectedAward = getSelectedAward;
 
         container.instance().componentDidMount();
         expect(getSelectedAward).toHaveBeenCalledTimes(1);
-        expect(getSelectedAward).toHaveBeenCalledWith(1234);
+        expect(getSelectedAward).toHaveBeenCalledWith('1234');
 
         const prevProps = Object.assign({}, mockParams, {
-            awardId: '222'
+            params: {
+                awardId: '222'
+            }
         });
 
         container.instance().componentDidUpdate(prevProps);
 
         expect(getSelectedAward).toHaveBeenCalledTimes(2);
-        expect(getSelectedAward).toHaveBeenLastCalledWith(1234);
+        expect(getSelectedAward).toHaveBeenLastCalledWith('1234');
     });
 
     describe('parseAward', () => {
         it('should parse returned contract data and send to the Redux store', () => {
-            const awardContainer = shallow(<AwardContainer
-                {...mockParams}
-                {...mockActions} />);
+            const awardContainer = getAwardContainer();
 
             const expectedAward = Object.create(BaseContract);
-            expectedAward.populate(mockApi);
+            expectedAward.populate(mockContract);
 
-            awardContainer.instance().parseAward(mockApi);
+            awardContainer.instance().parseAward(mockContract);
 
-            expect(mockActions.setSelectedAward).toHaveBeenCalledWith(expectedAward);
-        });
-        it('should parse returned financial assistance data and send to the Redux store', () => {
-            const awardContainer = shallow(<AwardContainer
-                {...mockParams}
-                {...mockActions} />);
-
-            const expectedAward = Object.create(BaseFinancialAssistance);
-            expectedAward.populate(mockFinancialAssistanceApi);
-
-            awardContainer.instance().parseAward(mockFinancialAssistanceApi);
-
-            expect(mockActions.setSelectedAward).toHaveBeenCalledWith(expectedAward);
+            expect(mockActions.setAward).toHaveBeenCalledWith(expectedAward);
         });
         it('should parse returned IDV data and send to the Redux store', () => {
-            const awardContainer = shallow(<AwardContainer
-                {...mockParams}
-                {...mockActions} />);
+            const awardContainer = getAwardContainer();
 
-            const mockIDV = Object.assign({}, mockApi, {
-                category: 'idv'
-            });
+            const expectedAward = Object.create(BaseIdv);
+            expectedAward.populate(mockIdv);
 
-            const expectedAward = Object.create(BaseContract);
-            expectedAward.populate(mockIDV);
+            awardContainer.instance().parseAward(mockIdv);
 
-            awardContainer.instance().parseAward(mockIDV);
+            expect(mockActions.setAward).toHaveBeenCalledWith(expectedAward);
+        });
+        it('should parse returned financial assistance data and send to the Redux store', () => {
+            const awardContainer = getAwardContainer();
 
-            expect(mockActions.setSelectedAward).toHaveBeenCalledWith(expectedAward);
+            const expectedAward = Object.create(BaseFinancialAssistance);
+            expectedAward.populate(mockLoan);
+
+            awardContainer.instance().parseAward(mockLoan);
+
+            expect(mockActions.setAward).toHaveBeenCalledWith(expectedAward);
+        });
+    });
+
+    describe('downloadData', () => {
+        const awardContainer = getAwardContainer();
+        const downloadRequest = {
+            promise: jest.fn(() => Promise.resolve({ results: { url: 'test', file_name: 'test.csv' } })),
+            cancel: jest.fn()
+        };
+
+        it("calls action-creator fns in props", async () => {
+            await awardContainer.instance().downloadData();
+            expect(mockActions.setDownloadCollapsed).toHaveBeenCalled();
+            expect(mockActions.setDownloadExpectedUrl).toHaveBeenCalled();
+            expect(mockActions.setDownloadExpectedFile).toHaveBeenCalled();
+            expect(mockActions.setDownloadPending).toHaveBeenCalled();
+        });
+
+        it("cancels the call if it's already pending", async () => {
+            awardContainer.instance().downloadRequest = downloadRequest;
+            await awardContainer.instance().downloadData();
+            expect(downloadRequest.cancel).toHaveBeenCalled();
+        });
+    });
+
+    describe('fetchAwardDownloadFile', () => {
+        it('returns the contract helper for contract awards', async () => {
+            const awardContainer = getAwardContainer();
+            let result = await awardContainer.instance().fetchAwardDownloadFile();
+            result = await result.promise;
+            expect(result.data.file_name).toEqual('contract.zip');
+        });
+
+        it('returns the idv helper for idv awards', async () => {
+            const awardContainer = getAwardContainer({ ...mockParams, award: { ...mockParams.award, category: "idv" } });
+            const result = await awardContainer.instance().fetchAwardDownloadFile().promise;
+            expect(result.data.file_name).toEqual('idv.zip');
+        });
+
+        it('returns the assistance helper for assistance awards', async () => {
+            const awardContainer = getAwardContainer({ ...mockParams, award: { ...mockParams.award, category: "grant" } });
+            const result = await awardContainer.instance().fetchAwardDownloadFile().promise;
+            expect(result.data.file_name).toEqual('assistance.zip');
         });
     });
 });
