+++ tests/containers/award/idv/IdvActivityContainer-test.jsx
@@ -6,15 +6,15 @@
 import React from 'react';
 import { mount, shallow } from 'enzyme';
 
-import { IdvActivityContainer } from 'containers/awardV2/idv/IdvActivityContainer';
+import { IdvActivityContainer } from 'containers/award/idv/IdvActivityContainer';
 import BaseIdvActivityBar from 'models/v2/awardsV2/BaseIdvActivityBar';
 import { mockRedux, mockActions } from '../mockAward';
 import { mockIdvActivity } from '../../../models/awardsV2/mockAwardApi';
 
-jest.mock('helpers/idvHelper', () => require('../awardV2Helper'));
+jest.mock('helpers/idvHelper', () => require('../awardHelper'));
 
 // mock the child component by replacing it with a function that returns a null element
-// jest.mock('components/awardv2/idv/amounts/AggregatedAwardAmounts.jsx', () => jest.fn(() => null));
+// jest.mock('components/award/idv/amounts/AggregatedAwardAmounts.jsx', () => jest.fn(() => null));
 
 describe('IdvActivityContainer', () => {
     const loadAwards = jest.fn();
