+++ tests/containers/award/mockAward.js
@@ -1,580 +1,39 @@
 /**
  * mockAward.js
- * Created by Kevin Li 3/9/17
+ * Created by David Trinh 10/6/18
  */
 
-export const mockAward = {
-    selectedAward: {
-        id: 57557,
-        award_id: 'DEAC0494AL85000',
-        recipient_name: 'SANDIA CORP',
-        description: 'MANAGEMENT AND OPERATION OF SANDIA NATIONAL LABORATORIES BY SANDIA CORPORATION, CONTRACT NO. DE-AC04-94AL85000',
-        date_signed: '11/30/2016',
-        period_of_performance_start_date: '11/30/2016',
-        period_of_performance_current_end_date: '4/30/2017',
-        award_type: 'D',
-        type: 'Unknown Type',
-        type_description: 'Unknown Type',
-        awarding_agency_name: 'Department of Energy',
-        awarding_subtier_name: 'Department of Energy',
-        funding_agency_name: 'Department of Defense',
-        funding_subtier_name: 'Department of the Air Force',
-        recipient_street: '1515 EUBANK BLVD. SEundefinednull',
-        recipient_city: 'ALBUQUERQUE',
-        recipient_state_province: 'IL',
-        recipient_zip_postal: '',
-        recipient_country: 'UNITED STATES',
-        parent_award: null,
-        pop_city_name: null,
-        pop_state_province: 'NM',
-        pop_zip: null,
-        total_obligation: '$88,074,906',
-        base_and_all_options_value: '$0',
-        recipient_duns: '007113228',
-        recipient_parent_duns: '',
-        recipient_business_type: 'Unknown Business Type',
-        type_of_contract_pricing: '',
-        type_of_contract_pricing_description: '',
-        latest_transaction: {
-            id: 40940,
-            data_source: 'DBR',
-            usaspending_unique_transaction_id: null,
-            type: 'D',
-            type_description: 'Unknown Type',
-            period_of_performance_start_date: '2016-11-30',
-            period_of_performance_current_end_date: '2017-04-30',
-            action_date: '2016-11-30',
-            action_type: 'C',
-            federal_action_obligation: '88074905.67',
-            modification_number: '634',
-            description: 'MANAGEMENT AND OPERATION OF SANDIA NATIONAL LABORATORIES BY SANDIA CORPORATION, CONTRACT NO. DE-AC04-94AL85000',
-            drv_award_transaction_usaspend: null,
-            drv_current_total_award_value_amount_adjustment: null,
-            drv_potential_total_award_value_amount_adjustment: null,
-            last_modified_date: null,
-            certified_date: null,
-            create_date: '2017-02-27T19:23:38.950068Z',
-            update_date: '2017-02-27T19:23:38.950104Z',
-            award: 57557,
-            submission: 9,
-            awarding_agency: {
-                id: 1611,
-                create_date: '2017-01-12T19:56:29.623000Z',
-                update_date: '2017-01-12T19:56:29.623000Z',
-                toptier_agency: {
-                    toptier_agency_id: 218,
-                    create_date: null,
-                    update_date: '2017-03-08T22:27:33.723892Z',
-                    toptier_code: '089',
-                    abbreviation: 'DOE',
-                    name: 'Department of Energy'
-                },
-                subtier_agency: {
-                    subtier_agency_id: 1477,
-                    create_date: null,
-                    update_date: '2017-03-08T22:27:29.546255Z',
-                    subtier_code: '8900',
-                    abbreviation: '',
-                    name: 'Department of Energy'
-                }
-            },
-            funding_agency: {
-                id: 1805,
-                create_date: '2017-01-12T19:56:30.604000Z',
-                update_date: '2017-01-12T19:56:30.604000Z',
-                toptier_agency: {
-                    toptier_agency_id: 268,
-                    create_date: null,
-                    update_date: '2017-03-08T22:27:44.239782Z',
-                    toptier_code: '097',
-                    abbreviation: 'DOD',
-                    name: 'Department of Defense'
-                },
-                subtier_agency: {
-                    subtier_agency_id: 1671,
-                    create_date: null,
-                    update_date: '2017-03-08T22:27:36.900322Z',
-                    subtier_code: '5700',
-                    abbreviation: '',
-                    name: 'Department of the Air Force'
-                }
-            },
-            recipient: {
-                legal_entity_id: 7113228,
-                data_source: 'DBR',
-                parent_recipient_unique_id: null,
-                recipient_name: 'SANDIA CORP',
-                vendor_doing_as_business_name: 'SANDIA NATIONAL LABORATORIES',
-                vendor_phone_number: null,
-                vendor_fax_number: null,
-                business_types: 'UN',
-                business_types_description: 'Unknown Business Type',
-                recipient_unique_id: '007113228',
-                limited_liability_corporation: '0',
-                sole_proprietorship: '0',
-                partnership_or_limited_liability_partnership: '0',
-                subchapter_scorporation: null,
-                foundation: '0',
-                for_profit_organization: '0',
-                nonprofit_organization: '0',
-                corporate_entity_tax_exempt: '0',
-                corporate_entity_not_tax_exempt: '0',
-                other_not_for_profit_organization: '0',
-                sam_exception: null,
-                city_local_government: '0',
-                county_local_government: '0',
-                inter_municipal_local_government: '0',
-                local_government_owned: '0',
-                municipality_local_government: '0',
-                school_district_local_government: '0',
-                township_local_government: '0',
-                us_state_government: '0',
-                us_federal_government: '0',
-                federal_agency: '0',
-                federally_funded_research_and_development_corp: '0',
-                us_tribal_government: '0',
-                foreign_government: '0',
-                community_developed_corporation_owned_firm: '0',
-                labor_surplus_area_firm: '0',
-                small_agricultural_cooperative: '0',
-                international_organization: '0',
-                us_government_entity: '0',
-                emerging_small_business: '0',
-                c8a_program_participant: '0',
-                sba_certified_8a_joint_venture: null,
-                dot_certified_disadvantage: '0',
-                self_certified_small_disadvantaged_business: '0',
-                historically_underutilized_business_zone: null,
-                small_disadvantaged_business: '0',
-                the_ability_one_program: '0',
-                historically_black_college: '0',
-                c1862_land_grant_college: '0',
-                c1890_land_grant_college: '0',
-                c1994_land_grant_college: '0',
-                minority_institution: '0',
-                private_university_or_college: '0',
-                school_of_forestry: '0',
-                state_controlled_institution_of_higher_learning: '0',
-                tribal_college: '0',
-                veterinary_college: '0',
-                higher_education: '0',
-                alaskan_native_servicing_institution: '0',
-                community_development_corporation: '0',
-                native_hawaiian_servicing_institution: '0',
-                domestic_shelter: '0',
-                manufacturer_of_goods: null,
-                hospital_flag: '0',
-                veterinary_hospital: '0',
-                hispanic_servicing_institution: '0',
-                woman_owned_business: '0',
-                minority_owned_business: '0',
-                women_owned_small_business: null,
-                economically_disadvantaged_women_owned_small_business: null,
-                joint_venture_women_owned_small_business: null,
-                joint_venture_economic_disadvantaged_women_owned_small_bus: null,
-                veteran_owned_business: '0',
-                service_disabled_veteran_owned_business: '0',
-                contracts: '0',
-                grants: '0',
-                receives_contracts_and_grants: '0',
-                airport_authority: '0',
-                council_of_governments: '0',
-                housing_authorities_public_tribal: '0',
-                interstate_entity: '0',
-                planning_commission: '0',
-                port_authority: '0',
-                transit_authority: '0',
-                foreign_owned_and_located: '0',
-                american_indian_owned_business: '0',
-                alaskan_native_owned_corporation_or_firm: '0',
-                indian_tribe_federally_recognized: '0',
-                native_hawaiian_owned_business: '0',
-                tribally_owned_business: '0',
-                asian_pacific_american_owned_business: '0',
-                black_american_owned_business: '0',
-                hispanic_american_owned_business: '0',
-                native_american_owned_business: '0',
-                subcontinent_asian_asian_indian_american_owned_business: '0',
-                other_minority_owned_business: '0',
-                us_local_government: '0',
-                undefinitized_action: 'X',
-                domestic_or_foreign_entity: null,
-                division_name: null,
-                division_number: null,
-                last_modified_date: null,
-                certified_date: null,
-                reporting_period_start: null,
-                reporting_period_end: null,
-                create_date: '2017-02-27T19:23:38.330219Z',
-                update_date: '2017-02-27T19:23:38.330243Z',
-                city_township_government: null,
-                special_district_government: null,
-                small_business: null,
-                individual: null,
-                location: {
-                    location_id: 54897,
-                    data_source: 'DBR',
-                    country_name: 'UNITED STATES',
-                    state_code: 'IL',
-                    state_name: null,
-                    state_description: null,
-                    city_name: 'ALBUQUERQUE',
-                    city_code: null,
-                    county_name: null,
-                    county_code: null,
-                    address_line1: '1515 EUBANK BLVD. SE',
-                    address_line2: 'P.O. BOX 5800, MS-0180, 87185-0180',
-                    address_line3: null,
-                    foreign_location_description: null,
-                    zip4: '871230180',
-                    zip_4a: null,
-                    congressional_code: null,
-                    performance_code: null,
-                    zip_last4: null,
-                    zip5: null,
-                    foreign_postal_code: null,
-                    foreign_province: 'NM',
-                    foreign_city_name: null,
-                    reporting_period_start: null,
-                    reporting_period_end: null,
-                    last_modified_date: null,
-                    certified_date: null,
-                    create_date: '2017-02-27T19:23:37.948162Z',
-                    update_date: '2017-02-27T19:23:37.948200Z',
-                    place_of_performance_flag: false,
-                    recipient_flag: true,
-                    location_country_code: 'USA'
-                }
-            },
-            place_of_performance: {
-                location_id: 54895,
-                data_source: 'DBR',
-                country_name: 'UNITED STATES',
-                state_code: 'NM',
-                state_name: null,
-                state_description: null,
-                city_name: null,
-                city_code: null,
-                county_name: null,
-                county_code: null,
-                address_line1: null,
-                address_line2: null,
-                address_line3: null,
-                foreign_location_description: null,
-                zip4: '871855400',
-                zip_4a: null,
-                congressional_code: '01',
-                performance_code: null,
-                zip_last4: null,
-                zip5: null,
-                foreign_postal_code: null,
-                foreign_province: 'NM',
-                foreign_city_name: null,
-                reporting_period_start: null,
-                reporting_period_end: null,
-                last_modified_date: null,
-                certified_date: null,
-                create_date: '2017-02-27T19:23:34.341138Z',
-                update_date: '2017-02-27T19:23:34.341162Z',
-                place_of_performance_flag: true,
-                recipient_flag: false,
-                location_country_code: 'USA'
-            },
-            assistance_data: null,
-            contract_data: {
-                transaction: 40940,
-                data_source: 'DBR',
-                piid: 'DEAC0494AL85000',
-                cost_or_pricing_data_description: null,
-                type_of_contract_pricing: 'R',
-                type_of_contract_pricing_description: null,
-                naics: null,
-                naics_description: null,
-                period_of_performance_potential_end_date: '2017-04-30',
-                ordering_period_end_date: null,
-                current_total_value_award: '0.00',
-                base_and_all_options_value: '0.00',
-                referenced_idv_agency_identifier: null,
-                idv_type: 'D',
-                multiple_or_single_award_idv: null,
-                type_of_idc: null,
-                a76_fair_act_action: null,
-                dod_claimant_program_code: null,
-                clinger_cohen_act_planning: '0',
-                commercial_item_acquisition_procedures_description: 'D',
-                commercial_item_test_program: '0',
-                consolidated_contract: '0',
-                contingency_humanitarian_or_peacekeeping_operation: null,
-                contract_bundling: null,
-                contract_financing: null,
-                contracting_officers_determination_of_business_size: 'O',
-                cost_accounting_standards: 'X',
-                country_of_product_or_service_origin: 'USA',
-                construction_wage_rat_desc: 'X',
-                evaluated_preference_description: null,
-                extent_competed_description: 'A',
-                fed_biz_opps_description: 'Y',
-                foreign_funding_description: 'X',
-                gfe_gfp: null,
-                information_technology_commercial_item_category_description: null,
-                interagency_contracting_authority_description: 'X',
-                local_area_set_aside: null,
-                major_program: null,
-                purchase_card_as_payment_method: '0',
-                multi_year_contract: '0',
-                national_interest_action: 'NONE',
-                number_of_actions: '1',
-                number_of_offers_received: '2',
-                other_statutory_authority: null,
-                performance_based_service_acquisition: 'Y',
-                place_of_manufacture: null,
-                price_evaluation_adjustment_preference_percent_difference: '0.00',
-                product_or_service_code: 'M181',
-                program_acronym: null,
-                other_than_full_and_open_competition: null,
-                recovered_materials_sustainability: 'B',
-                research: null,
-                sea_transportation_description: null,
-                labor_standards_description: 'N',
-                small_business_competitiveness_demonstration_program: null,
-                solicitation_identifier: null,
-                solicitation_procedures: null,
-                fair_opportunity_limited_sources_description: null,
-                subcontracting_plan_description: 'E',
-                program_system_or_equipment_code: null,
-                type_set_aside_description: null,
-                epa_designated_product: 'A',
-                materials_supplies_article: 'X',
-                transaction_number: '0',
-                referenced_idv_modification_number: null,
-                rec_flag: null,
-                drv_parent_award_awarding_agency_code: null,
-                drv_current_aggregated_total_value_of_award: null,
-                drv_current_total_value_of_award: null,
-                drv_potential_award_idv_amount_total_estimate: null,
-                drv_potential_aggregated_award_idv_amount_total_estimate: null,
-                drv_potential_aggregated_total_value_of_award: null,
-                drv_base_and_all_options_value: null,
-                create_date: '2017-02-27T19:23:39.321925Z',
-                update_date: '2017-02-27T19:23:39.321949Z',
-                last_modified_date: null,
-                certified_date: null,
-                reporting_period_start: '2016-11-01',
-                reporting_period_end: '2016-11-01',
-                submission: 9
-            }
-        }
+import { mockContract, mockIdv } from '../../models/awardsV2/mockAwardApi';
+
+export const mockParams = {
+    params: {
+        awardId: '1234'
     },
-    transactions: [
-        {
-            id: 40940,
-            type: 'D',
-            type_description: 'Unknown Type',
-            action_date: '11/30/2016',
-            action_type: 'C',
-            federal_action_obligation: '$88,074,906',
-            modification_number: '634',
-            description: 'MANAGEMENT AND OPERATION OF SANDIA NATIONAL LABORATORIES BY SANDIA CORPORATION, CONTRACT NO. DE-AC04-94AL85000'
-        }
-    ],
-    transactionMeta: {
-        count: 1,
-        page: 1,
-        totalPages: 1
-    },
-    renderHash: '5',
-    groupHash: '6',
-    transactionSort: {
-        field: 'modification_number',
-        direction: 'asc'
-    },
-    finSysData: [],
-    finSysMeta: {
-        count: 0,
-        page: 0,
-        totalPages: 0
-    },
-    finSysSort: {
-        field: 'certified_date',
-        direction: 'desc'
+    award: {
+        id: '1234',
+        category: 'contract',
+        overview: mockContract
     }
 };
 
-export const mockAwards = {
-    total_metadata: {
-        count: 2
-    },
-    page_metadata: {
-        page: 1,
-        has_next_page: true
-    },
-    results: [
-        {
-            financial_accounts_by_awards_id: 1,
-            program_activity_name: 'Reimbursable program activity',
-            certified_date: null,
-            treasury_account: {
-                treasury_account_identifier: 69481,
-                tas_rendering_label: '0892017/20180228',
-                account_title: 'Departmental Administration, Energy Programs, Energy',
-                reporting_agency_id: '089',
-                reporting_agency_name: 'Department of Energy'
-            },
-            program_activity_code: 801,
-            object_class: '254',
-            transaction_obligations: [{
-                transaction_obligated_amount: '-323015.00'
-            }]
-        }, {
-            financial_accounts_by_awards_id: 2,
-            program_activity_name: 'Wind Energy',
-            certified_date: null,
-            treasury_account: {
-                treasury_account_identifier: 35975,
-                tas_rendering_label: '089X0321',
-                account_title: 'Energy Efficiency and Renewable Energy, Energy Programs, Energy',
-                reporting_agency_id: '089',
-                reporting_agency_name: 'Department of Energy'
-            },
-            program_activity_code: 102,
-            object_class: '254',
-            transaction_obligations: [{
-                transaction_obligated_amount: '-11522.00'
-            }]
-        }
-    ]
+export const mockActions = {
+    setAward: jest.fn(),
+    setCounts: jest.fn(),
+    resetAward: jest.fn(),
+    setDownloadExpectedUrl: jest.fn(),
+    setDownloadExpectedFile: jest.fn(),
+    setDownloadPending: jest.fn(),
+    setDownloadCollapsed: jest.fn()
 };
 
-export const mockTransactions = [
-    {
-        id: 40904,
-        type: 'D',
-        type_description: 'Unknown Type',
-        action_date: '11/30/2016',
-        action_type: 'C',
-        federal_action_obligation: '$4,437,307',
-        modification_number: '959',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - TAS::97 4555::TAS - OBLIGATE SPP DOD FUNDING - NOVEMBER 30 2016'
-    },
-    {
-        id: 40905,
-        type: 'D',
-        type_description: 'Unknown Type',
-        action_date: '11/30/2016',
-        action_type: 'C',
-        federal_action_obligation: '$9,740,762',
-        modification_number: '958',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - OBLIGATE $9,740,762.00 IN DOE APPROPRIATED FUNDING AND $4,278,014.31 IN NON-APPROPRIATED FUNDS; CUMULATIVE NA FUNDS - $435,532,331.16.'
-    },
-    {
-        id: 40906,
-        type: 'D',
-        type_description: 'Unknown Type',
-        action_date: '11/17/2016',
-        action_type: 'C',
-        federal_action_obligation: '$241,000',
-        modification_number: '957',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - OBLIGATE $241,000.00 IN DOE APPROPRIATED FUNDING AND $0.00 IN NON-APPROPRIATED FUNDS; CUMULATIVE NA FUNDS - $431,254,316.85.'
-    },
-    {
-        id: 40907,
-        type: 'D',
-        type_description: 'Unknown Type',
-        action_date: '11/10/2016',
-        action_type: 'C',
-        federal_action_obligation: '$40,000',
-        modification_number: '956',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - OBLIGATE $40,000.00 IN DOE APPROPRIATED FUNDING AND $0.00 IN NON-APPROPRIATED FUNDS; CUMULATIVE NA FUNDS - $431,254,316.85.'
-    },
-    {
-        id: 40909,
-        type: 'D',
-        type_description: 'Unknown Type',
-        action_date: '11/7/2016',
-        action_type: 'C',
-        federal_action_obligation: '$21,007,166',
-        modification_number: '955',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - OBLIGATE $21,007,166.00 IN DOE APPROPRIATED FUNDING AND $0.00 IN NON-APPROPRIATED FUNDS; CUMULATIVE NA FUNDS - $431,254,316.85.'
-    },
-    {
-        id: 40908,
-        type: 'D',
-        type_description: 'Unknown Type',
-        action_date: '11/8/2016',
-        action_type: 'B',
-        federal_action_obligation: '$0',
-        modification_number: '954',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - REVISE APPENDIX E TO INCORPORATE RCN OR-66'
-    },
-    {
-        id: 39973,
-        type: 'D',
-        type_description: 'Definitive Contract',
-        action_date: '10/27/2016',
-        action_type: null,
-        federal_action_obligation: '$1,537,201',
-        modification_number: '953',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - TAS::97 0400::TAS - OBLIGATE SPP DOD FUNDING - OCTOBER 27 2016'
-    },
-    {
-        id: 39849,
-        type: 'D',
-        type_description: 'Definitive Contract',
-        action_date: '10/27/2016',
-        action_type: null,
-        federal_action_obligation: '$25,840,226',
-        modification_number: '952',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - OBLIGATE $25,840,225.75 IN DOE APPROPRIATED FUNDING AND $0.00 IN NON-APPROPRIATED FUNDS; CUMULATIVE NA FUNDS - $431,254,316.85.'
-    },
-    {
-        id: 39862,
-        type: 'D',
-        type_description: 'Definitive Contract',
-        action_date: '10/20/2016',
-        action_type: null,
-        federal_action_obligation: '$154,805,714',
-        modification_number: '951',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - OBLIGATE $154,805,714.00 IN DOE APPROPRIATED FUNDING AND $0.00 IN NON-APPROPRIATED FUNDS; CUMULATIVE NA FUNDS - $431,254,316.85.'
-    },
-    {
-        id: 36839,
-        type: 'D',
-        type_description: 'Definitive Contract',
-        action_date: '10/26/2015',
-        action_type: null,
-        federal_action_obligation: '$498,104',
-        modification_number: '913',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - TAS::68 0108::TAS - OBLIGATE SPP EPA FUNDING - OCTOBER 2015'
-    },
-    {
-        id: 36757,
-        type: 'D',
-        type_description: 'Definitive Contract',
-        action_date: '10/26/2015',
-        action_type: null,
-        federal_action_obligation: '$150,330,700',
-        modification_number: '912',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - OBLIGATE $150,330,700.26 IN DOE APPROPRIATED FUNDING; NON-APPROPRIATED FUNDS CURRENT ACTION = $1,523,053.45; NON-APPROPRIATED FUNDS CUMULATIVE SINCE MOD. 234 = $413,663,188.76.'
-    },
-    {
-        id: 36791,
-        type: 'D',
-        type_description: 'Definitive Contract',
-        action_date: '10/15/2015',
-        action_type: null,
-        federal_action_obligation: '$0',
-        modification_number: '911',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - REVISE APPENDIX E TO INCORPORATE RCN OR-62'
-    },
-    {
-        id: 34869,
-        type: 'D',
-        type_description: 'Definitive Contract',
-        action_date: '10/30/2014',
-        action_type: null,
-        federal_action_obligation: '-$14,274',
-        modification_number: '860',
-        description: 'MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY - RECOVERY TAS::89 0227::TAS - DEOBLIGATE SC ARRA FUNDING FROM PROJECT ENTITLED "NUCLEAR PHYSICS - NUCLEAR SCIENCE WORKFORCE"'
+export const mockRedux = {
+    award: {
+        id: '1234',
+        category: 'idv',
+        overview: mockIdv,
+        counts: {
+            idvs: 45,
+            contracts: 52
+        }
     }
-];
-
+};
