--- tests/containers/award/mockFinancialSystemDetails.js
@@ -1,36 +1,0 @@
-
-const items = {
-    total_metadata: {
-        count: 1
-    },
-    page_metadata: {
-        num_pages: 1,
-        page_number: 1,
-        count: 1
-    },
-    results: [
-        {
-            financial_accounts_by_awards_id: 22337,
-            program_activity_name: "Wind Energy",
-            certified_date: null,
-            submission: {
-                reporting_fiscal_year: '2017',
-                reporting_fiscal_quarter: '2'
-            },
-            treasury_account: {
-                treasury_account_identifier: 35975,
-                tas_rendering_label: "089X0321",
-                account_title: "Energy Efficiency and Renewable Energy, Energy Programs, Energy",
-                reporting_agency_id: "089",
-                reporting_agency_name: "Department of Energy"
-            },
-            program_activity_code: 102,
-            object_class: "254",
-            transaction_obligations: [{
-                transaction_obligated_amount: "-456163.00"
-            }]
-        }
-    ]
-};
-
-export default items;
