--- tests/containers/award/mockResults.js
@@ -1,1534 +1,0 @@
-export const mockApi = {
-    "id": 12178065,
-    "date_signed__fy": 2000,
-    "data_source": "DBR",
-    "type": "D",
-    "type_description": "DEFINITIVE CONTRACT",
-    "category": "contract",
-    "piid": "DEAC0500OR22725",
-    "fain": null,
-    "uri": null,
-    "total_obligation": "21491083751.90",
-    "total_outlay": null,
-    "total_subsidy_cost": null,
-    "date_signed": "1999-10-15",
-    "description": "MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY",
-    "period_of_performance_start_date": "1999-10-15",
-    "period_of_performance_current_end_date": "2020-03-31",
-    "potential_total_value_of_award": null,
-    "base_and_all_options_value": "21996778787.34",
-    "last_modified_date": "2018-02-23",
-    "certified_date": "2018-02-23",
-    "create_date": "2018-02-03T18:54:06.983782Z",
-    "update_date": "2018-03-01T01:57:23.050458Z",
-    "generated_unique_award_id": "CONT_AW_8900_-NONE-_DEAC0500OR22725_-NONE-",
-    "is_fpds": true,
-    "transaction_unique_id": "8900_-NONE-_DEAC0500OR22725_1033_-NONE-_0",
-    "total_subaward_amount": "0.00",
-    "subaward_count": 0,
-    "parent_award": null,
-    "awarding_agency": {
-        "id": 930,
-        "create_date": "2018-01-31T19:53:06.186372Z",
-        "update_date": "2018-03-07T19:54:58.010375Z",
-        "toptier_flag": true,
-        "toptier_agency": {
-            "toptier_agency_id": 78,
-            "create_date": "2018-01-31T19:53:06.173880Z",
-            "update_date": "2018-03-07T19:54:57.995826Z",
-            "toptier_code": "089",
-            "abbreviation": "DOE",
-            "name": "Department of Energy",
-            "mission": "The mission of the Energy Department is to ensure America's security and prosperity by addressing its energy, environmental and nuclear challenges through transformative science and technology solutions.",
-            "website": "https://energy.gov/",
-            "icon_filename": "DOEnergy.jpg"
-        },
-        "subtier_agency": {
-            "subtier_agency_id": 929,
-            "create_date": "2018-01-31T19:53:06.180583Z",
-            "update_date": "2018-03-07T19:54:58.003876Z",
-            "subtier_code": "8900",
-            "abbreviation": "DOE",
-            "name": "Department of Energy"
-        }
-    },
-    "funding_agency": {
-        "id": 930,
-        "create_date": "2018-01-31T19:53:06.186372Z",
-        "update_date": "2018-03-07T19:54:58.010375Z",
-        "toptier_flag": true,
-        "toptier_agency": {
-            "toptier_agency_id": 78,
-            "create_date": "2018-01-31T19:53:06.173880Z",
-            "update_date": "2018-03-07T19:54:57.995826Z",
-            "toptier_code": "089",
-            "abbreviation": "DOE",
-            "name": "Department of Energy",
-            "mission": "The mission of the Energy Department is to ensure America's security and prosperity by addressing its energy, environmental and nuclear challenges through transformative science and technology solutions.",
-            "website": "https://energy.gov/",
-            "icon_filename": "DOEnergy.jpg"
-        },
-        "subtier_agency": {
-            "subtier_agency_id": 929,
-            "create_date": "2018-01-31T19:53:06.180583Z",
-            "update_date": "2018-03-07T19:54:58.003876Z",
-            "subtier_code": "8900",
-            "abbreviation": "DOE",
-            "name": "Department of Energy"
-        }
-    },
-    "recipient": {
-        "legal_entity_id": 98140463,
-        "data_source": "DBR",
-        "parent_recipient_unique_id": "099114287",
-        "recipient_name": "UT BATTELLE LIMITED LIABILITY COMPANY",
-        "vendor_doing_as_business_name": "OAK RIDGE NATIONAL LABORATORY",
-        "vendor_phone_number": null,
-        "vendor_fax_number": null,
-        "business_types": null,
-        "business_types_description": null,
-        "business_categories": [
-            "minority_owned_business",
-            "nonprofit"
-        ],
-        "recipient_unique_id": "099114287",
-        "limited_liability_corporation": false,
-        "sole_proprietorship": false,
-        "partnership_or_limited_liability_partnership": false,
-        "subchapter_scorporation": false,
-        "foundation": false,
-        "for_profit_organization": false,
-        "nonprofit_organization": true,
-        "corporate_entity_tax_exempt": false,
-        "corporate_entity_not_tax_exempt": false,
-        "other_not_for_profit_organization": false,
-        "sam_exception": null,
-        "city_local_government": false,
-        "county_local_government": false,
-        "inter_municipal_local_government": false,
-        "local_government_owned": false,
-        "municipality_local_government": false,
-        "school_district_local_government": false,
-        "township_local_government": false,
-        "us_state_government": false,
-        "us_federal_government": false,
-        "federal_agency": false,
-        "federally_funded_research_and_development_corp": false,
-        "us_tribal_government": false,
-        "foreign_government": false,
-        "community_developed_corporation_owned_firm": false,
-        "labor_surplus_area_firm": false,
-        "small_agricultural_cooperative": false,
-        "international_organization": false,
-        "us_government_entity": false,
-        "emerging_small_business": false,
-        "c8a_program_participant": false,
-        "sba_certified_8a_joint_venture": false,
-        "dot_certified_disadvantage": false,
-        "self_certified_small_disadvantaged_business": false,
-        "historically_underutilized_business_zone": false,
-        "small_disadvantaged_business": false,
-        "the_ability_one_program": false,
-        "historically_black_college": false,
-        "c1862_land_grant_college": false,
-        "c1890_land_grant_college": false,
-        "c1994_land_grant_college": false,
-        "minority_institution": false,
-        "private_university_or_college": false,
-        "school_of_forestry": false,
-        "state_controlled_institution_of_higher_learning": false,
-        "tribal_college": false,
-        "veterinary_college": false,
-        "higher_education": true,
-        "alaskan_native_servicing_institution": false,
-        "community_development_corporation": false,
-        "native_hawaiian_servicing_institution": false,
-        "domestic_shelter": false,
-        "manufacturer_of_goods": false,
-        "hospital_flag": false,
-        "veterinary_hospital": false,
-        "hispanic_servicing_institution": false,
-        "woman_owned_business": false,
-        "minority_owned_business": true,
-        "women_owned_small_business": false,
-        "economically_disadvantaged_women_owned_small_business": false,
-        "joint_venture_women_owned_small_business": false,
-        "joint_venture_economic_disadvantaged_women_owned_small_bus": false,
-        "veteran_owned_business": false,
-        "service_disabled_veteran_owned_business": false,
-        "contracts": false,
-        "grants": false,
-        "receives_contracts_and_grants": false,
-        "airport_authority": false,
-        "council_of_governments": false,
-        "housing_authorities_public_tribal": false,
-        "interstate_entity": false,
-        "planning_commission": false,
-        "port_authority": false,
-        "transit_authority": false,
-        "foreign_owned_and_located": false,
-        "american_indian_owned_business": false,
-        "alaskan_native_owned_corporation_or_firm": false,
-        "indian_tribe_federally_recognized": false,
-        "native_hawaiian_owned_business": false,
-        "tribally_owned_business": false,
-        "asian_pacific_american_owned_business": false,
-        "black_american_owned_business": false,
-        "hispanic_american_owned_business": false,
-        "native_american_owned_business": false,
-        "subcontinent_asian_asian_indian_american_owned_business": false,
-        "other_minority_owned_business": false,
-        "us_local_government": false,
-        "undefinitized_action": "X",
-        "domestic_or_foreign_entity": null,
-        "domestic_or_foreign_entity_description": null,
-        "division_name": null,
-        "division_number": null,
-        "last_modified_date": null,
-        "certified_date": null,
-        "reporting_period_start": null,
-        "reporting_period_end": null,
-        "create_date": "2018-02-26T02:28:32.237016Z",
-        "update_date": "2018-02-26T02:28:32.238815Z",
-        "city_township_government": null,
-        "special_district_government": null,
-        "small_business": null,
-        "small_business_description": null,
-        "individual": null,
-        "is_fpds": true,
-        "transaction_unique_id": "NONE",
-        "location": {
-            "location_id": 196280925,
-            "data_source": "DBR",
-            "location_country_code": "USA",
-            "country_name": "UNITED STATES",
-            "state_code": "TN",
-            "state_name": "TENNESSEE",
-            "state_description": null,
-            "city_name": "OAK RIDGE",
-            "city_code": null,
-            "county_name": "ANDERSON",
-            "county_code": "001",
-            "address_line1": "BETHEL VALLEY ROAD",
-            "address_line2": null,
-            "address_line3": null,
-            "foreign_location_description": null,
-            "zip4": "378316396",
-            "zip_4a": null,
-            "congressional_code": null,
-            "performance_code": null,
-            "zip_last4": "6396",
-            "zip5": "37831",
-            "foreign_postal_code": null,
-            "foreign_province": null,
-            "foreign_city_name": null,
-            "reporting_period_start": null,
-            "reporting_period_end": null,
-            "last_modified_date": null,
-            "certified_date": null,
-            "create_date": "2018-02-26T02:28:32.235569Z",
-            "update_date": "2018-02-26T02:28:32.235587Z",
-            "place_of_performance_flag": false,
-            "recipient_flag": true,
-            "is_fpds": true,
-            "transaction_unique_id": "NONE"
-        }
-    },
-    "place_of_performance": {
-        "location_id": 196280926,
-        "data_source": "DBR",
-        "location_country_code": "USA",
-        "country_name": "UNITED STATES",
-        "state_code": "TN",
-        "state_name": "TENNESSEE",
-        "state_description": null,
-        "city_name": "OAK RIDGE",
-        "city_code": null,
-        "county_name": "ANDERSON",
-        "county_code": "001",
-        "address_line1": null,
-        "address_line2": null,
-        "address_line3": null,
-        "foreign_location_description": null,
-        "zip4": null,
-        "zip_4a": "378308313",
-        "congressional_code": "03",
-        "performance_code": null,
-        "zip_last4": "8313",
-        "zip5": "37830",
-        "foreign_postal_code": null,
-        "foreign_province": null,
-        "foreign_city_name": null,
-        "reporting_period_start": null,
-        "reporting_period_end": null,
-        "last_modified_date": null,
-        "certified_date": null,
-        "create_date": "2018-02-26T02:28:32.243738Z",
-        "update_date": "2018-02-26T02:28:32.243757Z",
-        "place_of_performance_flag": true,
-        "recipient_flag": false,
-        "is_fpds": false,
-        "transaction_unique_id": "NONE"
-    },
-    "latest_transaction": {
-        "id": 90223240,
-        "usaspending_unique_transaction_id": null,
-        "type": "D",
-        "type_description": "DEFINITIVE CONTRACT",
-        "period_of_performance_start_date": "2018-02-23",
-        "period_of_performance_current_end_date": "2020-03-31",
-        "action_date": "2018-02-23",
-        "action_type": "C",
-        "action_type_description": "FUNDING ONLY ACTION",
-        "federal_action_obligation": "8728668.18",
-        "original_loan_subsidy_cost": null,
-        "modification_number": "1038",
-        "description": "MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY TAS:: 80 0120 ::TAS SPP/NASA FUNDING  FEBRUARY 23, 2018",
-        "drv_award_transaction_usaspend": null,
-        "drv_current_total_award_value_amount_adjustment": null,
-        "drv_potential_total_award_value_amount_adjustment": null,
-        "last_modified_date": "2018-02-23",
-        "certified_date": null,
-        "create_date": "2018-02-24T21:24:35.377115Z",
-        "update_date": "2018-02-26T02:28:32.260686Z",
-        "fiscal_year": 2018,
-        "transaction_unique_id": "8900_-NONE-_DEAC0500OR22725_1038_-NONE-_0",
-        "generated_unique_award_id": "CONT_AW_8900_-NONE-_DEAC0500OR22725_-NONE-",
-        "is_fpds": true,
-        "award": 12178065,
-        "awarding_agency": {
-            "id": 930,
-            "create_date": "2018-01-31T19:53:06.186372Z",
-            "update_date": "2018-03-07T19:54:58.010375Z",
-            "toptier_flag": true,
-            "toptier_agency": {
-                "toptier_agency_id": 78,
-                "create_date": "2018-01-31T19:53:06.173880Z",
-                "update_date": "2018-03-07T19:54:57.995826Z",
-                "toptier_code": "089",
-                "abbreviation": "DOE",
-                "name": "Department of Energy",
-                "mission": "The mission of the Energy Department is to ensure America's security and prosperity by addressing its energy, environmental and nuclear challenges through transformative science and technology solutions.",
-                "website": "https://energy.gov/",
-                "icon_filename": "DOEnergy.jpg"
-            },
-            "subtier_agency": {
-                "subtier_agency_id": 929,
-                "create_date": "2018-01-31T19:53:06.180583Z",
-                "update_date": "2018-03-07T19:54:58.003876Z",
-                "subtier_code": "8900",
-                "abbreviation": "DOE",
-                "name": "Department of Energy"
-            }
-        },
-        "funding_agency": {
-            "id": 930,
-            "create_date": "2018-01-31T19:53:06.186372Z",
-            "update_date": "2018-03-07T19:54:58.010375Z",
-            "toptier_flag": true,
-            "toptier_agency": {
-                "toptier_agency_id": 78,
-                "create_date": "2018-01-31T19:53:06.173880Z",
-                "update_date": "2018-03-07T19:54:57.995826Z",
-                "toptier_code": "089",
-                "abbreviation": "DOE",
-                "name": "Department of Energy",
-                "mission": "The mission of the Energy Department is to ensure America's security and prosperity by addressing its energy, environmental and nuclear challenges through transformative science and technology solutions.",
-                "website": "https://energy.gov/",
-                "icon_filename": "DOEnergy.jpg"
-            },
-            "subtier_agency": {
-                "subtier_agency_id": 929,
-                "create_date": "2018-01-31T19:53:06.180583Z",
-                "update_date": "2018-03-07T19:54:58.003876Z",
-                "subtier_code": "8900",
-                "abbreviation": "DOE",
-                "name": "Department of Energy"
-            }
-        },
-        "recipient": {
-            "legal_entity_id": 98140463,
-            "data_source": "DBR",
-            "parent_recipient_unique_id": "099114287",
-            "recipient_name": "UT BATTELLE LIMITED LIABILITY COMPANY",
-            "vendor_doing_as_business_name": "OAK RIDGE NATIONAL LABORATORY",
-            "vendor_phone_number": null,
-            "vendor_fax_number": null,
-            "business_types": null,
-            "business_types_description": null,
-            "business_categories": [
-                "minority_owned_business",
-                "nonprofit"
-            ],
-            "recipient_unique_id": "099114287",
-            "limited_liability_corporation": false,
-            "sole_proprietorship": false,
-            "partnership_or_limited_liability_partnership": false,
-            "subchapter_scorporation": false,
-            "foundation": false,
-            "for_profit_organization": false,
-            "nonprofit_organization": true,
-            "corporate_entity_tax_exempt": false,
-            "corporate_entity_not_tax_exempt": false,
-            "other_not_for_profit_organization": false,
-            "sam_exception": null,
-            "city_local_government": false,
-            "county_local_government": false,
-            "inter_municipal_local_government": false,
-            "local_government_owned": false,
-            "municipality_local_government": false,
-            "school_district_local_government": false,
-            "township_local_government": false,
-            "us_state_government": false,
-            "us_federal_government": false,
-            "federal_agency": false,
-            "federally_funded_research_and_development_corp": false,
-            "us_tribal_government": false,
-            "foreign_government": false,
-            "community_developed_corporation_owned_firm": false,
-            "labor_surplus_area_firm": false,
-            "small_agricultural_cooperative": false,
-            "international_organization": false,
-            "us_government_entity": false,
-            "emerging_small_business": false,
-            "c8a_program_participant": false,
-            "sba_certified_8a_joint_venture": false,
-            "dot_certified_disadvantage": false,
-            "self_certified_small_disadvantaged_business": false,
-            "historically_underutilized_business_zone": false,
-            "small_disadvantaged_business": false,
-            "the_ability_one_program": false,
-            "historically_black_college": false,
-            "c1862_land_grant_college": false,
-            "c1890_land_grant_college": false,
-            "c1994_land_grant_college": false,
-            "minority_institution": false,
-            "private_university_or_college": false,
-            "school_of_forestry": false,
-            "state_controlled_institution_of_higher_learning": false,
-            "tribal_college": false,
-            "veterinary_college": false,
-            "higher_education": true,
-            "alaskan_native_servicing_institution": false,
-            "community_development_corporation": false,
-            "native_hawaiian_servicing_institution": false,
-            "domestic_shelter": false,
-            "manufacturer_of_goods": false,
-            "hospital_flag": false,
-            "veterinary_hospital": false,
-            "hispanic_servicing_institution": false,
-            "woman_owned_business": false,
-            "minority_owned_business": true,
-            "women_owned_small_business": false,
-            "economically_disadvantaged_women_owned_small_business": false,
-            "joint_venture_women_owned_small_business": false,
-            "joint_venture_economic_disadvantaged_women_owned_small_bus": false,
-            "veteran_owned_business": false,
-            "service_disabled_veteran_owned_business": false,
-            "contracts": false,
-            "grants": false,
-            "receives_contracts_and_grants": false,
-            "airport_authority": false,
-            "council_of_governments": false,
-            "housing_authorities_public_tribal": false,
-            "interstate_entity": false,
-            "planning_commission": false,
-            "port_authority": false,
-            "transit_authority": false,
-            "foreign_owned_and_located": false,
-            "american_indian_owned_business": false,
-            "alaskan_native_owned_corporation_or_firm": false,
-            "indian_tribe_federally_recognized": false,
-            "native_hawaiian_owned_business": false,
-            "tribally_owned_business": false,
-            "asian_pacific_american_owned_business": false,
-            "black_american_owned_business": false,
-            "hispanic_american_owned_business": false,
-            "native_american_owned_business": false,
-            "subcontinent_asian_asian_indian_american_owned_business": false,
-            "other_minority_owned_business": false,
-            "us_local_government": false,
-            "undefinitized_action": "X",
-            "domestic_or_foreign_entity": null,
-            "domestic_or_foreign_entity_description": null,
-            "division_name": null,
-            "division_number": null,
-            "last_modified_date": null,
-            "certified_date": null,
-            "reporting_period_start": null,
-            "reporting_period_end": null,
-            "create_date": "2018-02-26T02:28:32.237016Z",
-            "update_date": "2018-02-26T02:28:32.238815Z",
-            "city_township_government": null,
-            "special_district_government": null,
-            "small_business": null,
-            "small_business_description": null,
-            "individual": null,
-            "is_fpds": true,
-            "transaction_unique_id": "NONE",
-            "location": {
-                "location_id": 196280925,
-                "data_source": "DBR",
-                "location_country_code": "USA",
-                "country_name": "UNITED STATES",
-                "state_code": "TN",
-                "state_name": "TENNESSEE",
-                "state_description": null,
-                "city_name": "OAK RIDGE",
-                "city_code": null,
-                "county_name": "ANDERSON",
-                "county_code": "001",
-                "address_line1": "BETHEL VALLEY ROAD",
-                "address_line2": null,
-                "address_line3": null,
-                "foreign_location_description": null,
-                "zip4": "378316396",
-                "zip_4a": null,
-                "congressional_code": null,
-                "performance_code": null,
-                "zip_last4": "6396",
-                "zip5": "37831",
-                "foreign_postal_code": null,
-                "foreign_province": null,
-                "foreign_city_name": null,
-                "reporting_period_start": null,
-                "reporting_period_end": null,
-                "last_modified_date": null,
-                "certified_date": null,
-                "create_date": "2018-02-26T02:28:32.235569Z",
-                "update_date": "2018-02-26T02:28:32.235587Z",
-                "place_of_performance_flag": false,
-                "recipient_flag": true,
-                "is_fpds": true,
-                "transaction_unique_id": "NONE"
-            }
-        },
-        "place_of_performance": {
-            "location_id": 196280926,
-            "data_source": "DBR",
-            "location_country_code": "USA",
-            "country_name": "UNITED STATES",
-            "state_code": "TN",
-            "state_name": "TENNESSEE",
-            "state_description": null,
-            "city_name": "OAK RIDGE",
-            "city_code": null,
-            "county_name": "ANDERSON",
-            "county_code": "001",
-            "address_line1": null,
-            "address_line2": null,
-            "address_line3": null,
-            "foreign_location_description": null,
-            "zip4": null,
-            "zip_4a": "378308313",
-            "congressional_code": "03",
-            "performance_code": null,
-            "zip_last4": "8313",
-            "zip5": "37830",
-            "foreign_postal_code": null,
-            "foreign_province": null,
-            "foreign_city_name": null,
-            "reporting_period_start": null,
-            "reporting_period_end": null,
-            "last_modified_date": null,
-            "certified_date": null,
-            "create_date": "2018-02-26T02:28:32.243738Z",
-            "update_date": "2018-02-26T02:28:32.243757Z",
-            "place_of_performance_flag": true,
-            "recipient_flag": false,
-            "is_fpds": false,
-            "transaction_unique_id": "NONE"
-        },
-        "assistance_data": null,
-        "contract_data": {
-            "transaction": 90223240,
-            "detached_award_procurement_id": "78671000",
-            "detached_award_proc_unique": "8900_-NONE-_DEAC0500OR22725_1038_-NONE-_0",
-            "piid": "DEAC0500OR22725",
-            "agency_id": "8900",
-            "awarding_sub_tier_agency_c": "8900",
-            "awarding_sub_tier_agency_n": "ENERGY, DEPARTMENT OF",
-            "awarding_agency_code": "089",
-            "awarding_agency_name": "DEPARTMENT OF ENERGY (DOE)",
-            "parent_award_id": null,
-            "award_modification_amendme": "1038",
-            "type_of_contract_pricing": "R",
-            "type_of_contract_pric_desc": "COST PLUS AWARD FEE",
-            "contract_award_type": "D",
-            "contract_award_type_desc": "DEFINITIVE CONTRACT",
-            "naics": "561210",
-            "naics_description": "FACILITIES SUPPORT SERVICES",
-            "awardee_or_recipient_uniqu": "099114287",
-            "ultimate_parent_legal_enti": "UT-BATTELLE  LLC",
-            "ultimate_parent_unique_ide": "099114287",
-            "award_description": "MANAGEMENT AND OPERATION OF THE OAK RIDGE NATIONAL LABORATORY TAS:: 80 0120 ::TAS SPP/NASA FUNDING  FEBRUARY 23, 2018",
-            "place_of_performance_zip4a": "378308313",
-            "place_of_performance_zip5": "37830",
-            "place_of_perform_zip_last4": "8313",
-            "place_of_perform_city_name": "OAK RIDGE",
-            "place_of_perform_county_na": "ANDERSON",
-            "place_of_perform_county_co": "001",
-            "place_of_performance_congr": "03",
-            "awardee_or_recipient_legal": "UT BATTELLE LIMITED LIABILITY COMPANY",
-            "legal_entity_city_name": "OAK RIDGE",
-            "legal_entity_state_code": "TN",
-            "legal_entity_state_descrip": "TENNESSEE",
-            "legal_entity_county_code": "001",
-            "legal_entity_county_name": "ANDERSON",
-            "legal_entity_zip4": "378316396",
-            "legal_entity_zip5": "37831",
-            "legal_entity_zip_last4": "6396",
-            "legal_entity_congressional": null,
-            "legal_entity_address_line1": "BETHEL VALLEY ROAD",
-            "legal_entity_address_line2": null,
-            "legal_entity_address_line3": null,
-            "legal_entity_country_code": "USA",
-            "legal_entity_country_name": "UNITED STATES",
-            "period_of_performance_star": "2018-02-23 00:00:00",
-            "period_of_performance_curr": "2020-03-31 00:00:00",
-            "period_of_perf_potential_e": "2020-03-31 00:00:00",
-            "ordering_period_end_date": null,
-            "action_date": "2018-02-23",
-            "action_type": "C",
-            "action_type_description": "FUNDING ONLY ACTION",
-            "federal_action_obligation": "8728668.18",
-            "current_total_value_award": "21996778787.34",
-            "potential_total_value_awar": "21996778787.34",
-            "base_exercised_options_val": "8728668.18",
-            "base_and_all_options_value": "8728668.18",
-            "funding_sub_tier_agency_co": "8900",
-            "funding_sub_tier_agency_na": "ENERGY, DEPARTMENT OF",
-            "funding_office_code": "892431",
-            "funding_office_name": "SC OAK RIDGE OFFICE",
-            "awarding_office_code": "892431",
-            "awarding_office_name": "SC OAK RIDGE OFFICE",
-            "referenced_idv_agency_iden": null,
-            "referenced_idv_agency_desc": null,
-            "funding_agency_code": "089",
-            "funding_agency_name": "DEPARTMENT OF ENERGY (DOE)",
-            "place_of_performance_locat": null,
-            "place_of_performance_state": "TN",
-            "place_of_perfor_state_desc": "TENNESSEE",
-            "place_of_perform_country_c": "USA",
-            "place_of_perf_country_desc": "UNITED STATES",
-            "idv_type": null,
-            "idv_type_description": null,
-            "award_or_idv_flag": null,
-            "referenced_idv_type": null,
-            "referenced_idv_type_desc": null,
-            "vendor_doing_as_business_n": "OAK RIDGE NATIONAL LABORATORY",
-            "vendor_phone_number": null,
-            "vendor_fax_number": null,
-            "multiple_or_single_award_i": null,
-            "multiple_or_single_aw_desc": null,
-            "referenced_mult_or_single": null,
-            "referenced_mult_or_si_desc": null,
-            "type_of_idc": null,
-            "type_of_idc_description": null,
-            "a_76_fair_act_action": "N",
-            "a_76_fair_act_action_desc": "NO",
-            "dod_claimant_program_code": null,
-            "dod_claimant_prog_cod_desc": null,
-            "clinger_cohen_act_planning": "N",
-            "clinger_cohen_act_pla_desc": "NO",
-            "commercial_item_acquisitio": "D",
-            "commercial_item_acqui_desc": "COMMERCIAL ITEM PROCEDURES NOT USED",
-            "commercial_item_test_progr": "N",
-            "commercial_item_test_desc": "NO",
-            "consolidated_contract": "N",
-            "consolidated_contract_desc": "NO",
-            "contingency_humanitarian_o": "X",
-            "contingency_humanitar_desc": "NOT APPLICABLE",
-            "contract_bundling": "D",
-            "contract_bundling_descrip": "NOT A BUNDLED REQUIREMENT",
-            "contract_financing": null,
-            "contract_financing_descrip": null,
-            "contracting_officers_deter": "O",
-            "contracting_officers_desc": "OTHER THAN SMALL BUSINESS",
-            "cost_accounting_standards": "X",
-            "cost_accounting_stand_desc": "NOT APPLICABLE EXEMPT FROM CAS",
-            "cost_or_pricing_data": null,
-            "cost_or_pricing_data_desc": null,
-            "country_of_product_or_serv": "USA",
-            "country_of_product_or_desc": "UNITED STATES",
-            "construction_wage_rate_req": "X",
-            "construction_wage_rat_desc": "NOT APPLICABLE",
-            "evaluated_preference": null,
-            "evaluated_preference_desc": null,
-            "extent_competed": "A",
-            "extent_compete_description": "FULL AND OPEN COMPETITION",
-            "fed_biz_opps": "Y",
-            "fed_biz_opps_description": "YES",
-            "foreign_funding": "X",
-            "foreign_funding_desc": "NOT APPLICABLE",
-            "government_furnished_prope": "N",
-            "government_furnished_desc": "TRANSACTION DOES NOT USE GFE/GFP",
-            "information_technology_com": null,
-            "information_technolog_desc": null,
-            "interagency_contracting_au": "A",
-            "interagency_contract_desc": "ECONOMY ACT",
-            "local_area_set_aside": null,
-            "local_area_set_aside_desc": null,
-            "major_program": null,
-            "purchase_card_as_payment_m": "N",
-            "purchase_card_as_paym_desc": "NO",
-            "multi_year_contract": "N",
-            "multi_year_contract_desc": "NO",
-            "national_interest_action": "NONE",
-            "national_interest_desc": "NONE",
-            "number_of_actions": "1",
-            "number_of_offers_received": "2",
-            "other_statutory_authority": null,
-            "performance_based_service": "N",
-            "performance_based_se_desc": "NO - SERVICE WHERE PBA IS NOT USED.",
-            "place_of_manufacture": "C",
-            "place_of_manufacture_desc": "NOT A MANUFACTURED END PRODUCT",
-            "price_evaluation_adjustmen": "0.00",
-            "product_or_service_code": "M181",
-            "product_or_service_co_desc": "OPER OF GOVT R&D GOCO FACILITIES",
-            "program_acronym": null,
-            "other_than_full_and_open_c": null,
-            "other_than_full_and_o_desc": null,
-            "recovered_materials_sustai": "C",
-            "recovered_materials_s_desc": "NO CLAUSES INCLUDED AND NO SUSTAINABILITY INCLUDED",
-            "research": null,
-            "research_description": null,
-            "sea_transportation": null,
-            "sea_transportation_desc": null,
-            "labor_standards": "X",
-            "labor_standards_descrip": "NOT APPLICABLE",
-            "small_business_competitive": "False",
-            "solicitation_identifier": null,
-            "solicitation_procedures": null,
-            "solicitation_procedur_desc": null,
-            "fair_opportunity_limited_s": null,
-            "fair_opportunity_limi_desc": null,
-            "subcontracting_plan": "C",
-            "subcontracting_plan_desc": "PLAN REQUIRED - INCENTIVE NOT INCLUDED ",
-            "program_system_or_equipmen": null,
-            "program_system_or_equ_desc": null,
-            "type_set_aside": null,
-            "type_set_aside_description": null,
-            "epa_designated_product": null,
-            "epa_designated_produc_desc": null,
-            "materials_supplies_article": "Y",
-            "materials_supplies_descrip": "YES",
-            "transaction_number": "0",
-            "sam_exception": null,
-            "sam_exception_description": null,
-            "city_local_government": "False",
-            "county_local_government": "False",
-            "inter_municipal_local_gove": "False",
-            "local_government_owned": "False",
-            "municipality_local_governm": "False",
-            "school_district_local_gove": "False",
-            "township_local_government": "False",
-            "us_state_government": "False",
-            "us_federal_government": "False",
-            "federal_agency": "False",
-            "federally_funded_research": "False",
-            "us_tribal_government": "False",
-            "foreign_government": "False",
-            "community_developed_corpor": "False",
-            "labor_surplus_area_firm": "False",
-            "corporate_entity_not_tax_e": "False",
-            "corporate_entity_tax_exemp": "False",
-            "partnership_or_limited_lia": "False",
-            "sole_proprietorship": "False",
-            "small_agricultural_coopera": "False",
-            "international_organization": "False",
-            "us_government_entity": "False",
-            "emerging_small_business": "False",
-            "c8a_program_participant": "False",
-            "sba_certified_8_a_joint_ve": "False",
-            "dot_certified_disadvantage": "False",
-            "self_certified_small_disad": "False",
-            "historically_underutilized": "False",
-            "small_disadvantaged_busine": "False",
-            "the_ability_one_program": "False",
-            "historically_black_college": "False",
-            "c1862_land_grant_college": "False",
-            "c1890_land_grant_college": "False",
-            "c1994_land_grant_college": "False",
-            "minority_institution": "False",
-            "private_university_or_coll": "False",
-            "school_of_forestry": "False",
-            "state_controlled_instituti": "False",
-            "tribal_college": "False",
-            "veterinary_college": "False",
-            "higher_education": "True",
-            "alaskan_native_servicing_i": "False",
-            "community_development_corp": "False",
-            "native_hawaiian_servicing": "False",
-            "domestic_shelter": "False",
-            "manufacturer_of_goods": "False",
-            "hospital_flag": "False",
-            "veterinary_hospital": "False",
-            "hispanic_servicing_institu": "False",
-            "foundation": "False",
-            "woman_owned_business": "False",
-            "minority_owned_business": "True",
-            "women_owned_small_business": "False",
-            "economically_disadvantaged": "False",
-            "joint_venture_women_owned": "False",
-            "joint_venture_economically": "False",
-            "veteran_owned_business": "False",
-            "service_disabled_veteran_o": "False",
-            "contracts": "False",
-            "grants": "False",
-            "receives_contracts_and_gra": "False",
-            "airport_authority": "False",
-            "council_of_governments": "False",
-            "housing_authorities_public": "False",
-            "interstate_entity": "False",
-            "planning_commission": "False",
-            "port_authority": "False",
-            "transit_authority": "False",
-            "subchapter_s_corporation": "False",
-            "limited_liability_corporat": "False",
-            "foreign_owned_and_located": "False",
-            "american_indian_owned_busi": "False",
-            "alaskan_native_owned_corpo": "False",
-            "indian_tribe_federally_rec": "False",
-            "native_hawaiian_owned_busi": "False",
-            "tribally_owned_business": "False",
-            "asian_pacific_american_own": "False",
-            "black_american_owned_busin": "False",
-            "hispanic_american_owned_bu": "False",
-            "native_american_owned_busi": "False",
-            "subcontinent_asian_asian_i": "False",
-            "other_minority_owned_busin": "False",
-            "for_profit_organization": "False",
-            "nonprofit_organization": "True",
-            "other_not_for_profit_organ": "False",
-            "us_local_government": "False",
-            "referenced_idv_modificatio": null,
-            "undefinitized_action": "X",
-            "undefinitized_action_desc": "NO",
-            "domestic_or_foreign_entity": null,
-            "domestic_or_foreign_e_desc": null,
-            "annual_revenue": "964000000",
-            "division_name": null,
-            "division_number_or_office": null,
-            "number_of_employees": "4000",
-            "vendor_alternate_name": "OAK RIDGE NATIONAL LABORATORY",
-            "vendor_alternate_site_code": "378316396",
-            "vendor_enabled": null,
-            "vendor_legal_org_name": "UT BATTELLE LIMITED LIABILITY COMPANY",
-            "vendor_location_disabled_f": null,
-            "vendor_site_code": "0991142870000PR",
-            "pulled_from": "AWARD",
-            "last_modified": "2018-02-23 15:06:48",
-            "initial_report_date": "2018-02-23",
-            "created_at": "2018-02-24T11:43:43.257667",
-            "updated_at": "2018-02-25T11:36:45.751639",
-            "officer_1_name": "Thomas E Mason",
-            "officer_1_amount": "905226.00",
-            "officer_2_name": "Jeffrey W Smith",
-            "officer_2_amount": "713363.00",
-            "officer_3_name": "Michelle V Buchanan",
-            "officer_3_amount": "584927.00",
-            "officer_4_name": "Alan S Icenhour",
-            "officer_4_amount": "576762.00",
-            "officer_5_name": "James B Roberto",
-            "officer_5_amount": "573902.00"
-        }
-    }
-};
-
-export const mockFinancialAssistanceApi = {
-    "id": 44728165,
-    "date_signed__fy": 2018,
-    "data_source": "DBR",
-    "type": "03",
-    "type_description": null,
-    "category": "grant",
-    "piid": null,
-    "fain": "1805NY5MAP",
-    "uri": null,
-    "total_obligation": "20196062000.00",
-    "total_outlay": null,
-    "total_subsidy_cost": "0.00",
-    "date_signed": "2017-10-01",
-    "description": "MEDICAID ENTITLEMENT FOR 42 - FY 2018 QUARTER 1 - T19",
-    "period_of_performance_start_date": "2017-10-01",
-    "period_of_performance_current_end_date": "2017-12-31",
-    "potential_total_value_of_award": null,
-    "base_and_all_options_value": null,
-    "last_modified_date": "2018-01-19",
-    "certified_date": "2018-01-03",
-    "create_date": "2018-02-03T18:54:06.983782Z",
-    "update_date": "2018-02-03T18:54:06.983782Z",
-    "generated_unique_award_id": "ASST_AW_7530_1805NY5MAP_-NONE-",
-    "is_fpds": false,
-    "transaction_unique_id": "5_7530_1805NY5MAP_7530-227-1805NY5MAP-5-1-2018-93778-75-0512-NON",
-    "total_subaward_amount": "0.00",
-    "subaward_count": 0,
-    "parent_award": null,
-    "awarding_agency": {
-        "id": 831,
-        "create_date": "2018-01-31T19:53:04.882146Z",
-        "update_date": "2018-03-07T19:54:56.491927Z",
-        "toptier_flag": false,
-        "toptier_agency": {
-            "toptier_agency_id": 68,
-            "create_date": "2018-01-31T19:53:04.542546Z",
-            "update_date": "2018-03-07T19:54:56.115928Z",
-            "toptier_code": "075",
-            "abbreviation": "HHS",
-            "name": "Department of Health and Human Services",
-            "mission": "It is the mission of the U.S. Department of Health & Human Services (HHS) to enhance and protect the health and well-being of all Americans. We fulfill that mission by providing for effective health and human services and fostering advances in medicine, public health, and social services.",
-            "website": "https://www.hhs.gov/",
-            "icon_filename": "DHHS.jpg"
-        },
-        "subtier_agency": {
-            "subtier_agency_id": 830,
-            "create_date": "2018-01-31T19:53:04.876355Z",
-            "update_date": "2018-03-07T19:54:56.485524Z",
-            "subtier_code": "7530",
-            "abbreviation": "CMS",
-            "name": "Centers for Medicare and Medicaid Services"
-        }
-    },
-    "funding_agency": null,
-    "recipient": {
-        "legal_entity_id": 83053945,
-        "data_source": "DBR",
-        "parent_recipient_unique_id": null,
-        "recipient_name": "NEW YORK STATE DEPARTMENT OF HEALTH",
-        "vendor_doing_as_business_name": null,
-        "vendor_phone_number": null,
-        "vendor_fax_number": null,
-        "business_types": "A",
-        "business_types_description": "STATE GOVERNMENT",
-        "business_categories": [
-            "regional_and_state_government",
-            "government"
-        ],
-        "recipient_unique_id": "806781340",
-        "limited_liability_corporation": null,
-        "sole_proprietorship": null,
-        "partnership_or_limited_liability_partnership": null,
-        "subchapter_scorporation": null,
-        "foundation": null,
-        "for_profit_organization": null,
-        "nonprofit_organization": null,
-        "corporate_entity_tax_exempt": null,
-        "corporate_entity_not_tax_exempt": null,
-        "other_not_for_profit_organization": null,
-        "sam_exception": null,
-        "city_local_government": null,
-        "county_local_government": null,
-        "inter_municipal_local_government": null,
-        "local_government_owned": null,
-        "municipality_local_government": null,
-        "school_district_local_government": null,
-        "township_local_government": null,
-        "us_state_government": null,
-        "us_federal_government": null,
-        "federal_agency": null,
-        "federally_funded_research_and_development_corp": null,
-        "us_tribal_government": null,
-        "foreign_government": null,
-        "community_developed_corporation_owned_firm": null,
-        "labor_surplus_area_firm": null,
-        "small_agricultural_cooperative": null,
-        "international_organization": null,
-        "us_government_entity": null,
-        "emerging_small_business": null,
-        "c8a_program_participant": null,
-        "sba_certified_8a_joint_venture": null,
-        "dot_certified_disadvantage": null,
-        "self_certified_small_disadvantaged_business": null,
-        "historically_underutilized_business_zone": null,
-        "small_disadvantaged_business": null,
-        "the_ability_one_program": null,
-        "historically_black_college": null,
-        "c1862_land_grant_college": null,
-        "c1890_land_grant_college": null,
-        "c1994_land_grant_college": null,
-        "minority_institution": null,
-        "private_university_or_college": null,
-        "school_of_forestry": null,
-        "state_controlled_institution_of_higher_learning": null,
-        "tribal_college": null,
-        "veterinary_college": null,
-        "higher_education": null,
-        "alaskan_native_servicing_institution": null,
-        "community_development_corporation": null,
-        "native_hawaiian_servicing_institution": null,
-        "domestic_shelter": null,
-        "manufacturer_of_goods": null,
-        "hospital_flag": null,
-        "veterinary_hospital": null,
-        "hispanic_servicing_institution": null,
-        "woman_owned_business": null,
-        "minority_owned_business": null,
-        "women_owned_small_business": null,
-        "economically_disadvantaged_women_owned_small_business": null,
-        "joint_venture_women_owned_small_business": null,
-        "joint_venture_economic_disadvantaged_women_owned_small_bus": null,
-        "veteran_owned_business": null,
-        "service_disabled_veteran_owned_business": null,
-        "contracts": null,
-        "grants": null,
-        "receives_contracts_and_grants": null,
-        "airport_authority": null,
-        "council_of_governments": null,
-        "housing_authorities_public_tribal": null,
-        "interstate_entity": null,
-        "planning_commission": null,
-        "port_authority": null,
-        "transit_authority": null,
-        "foreign_owned_and_located": null,
-        "american_indian_owned_business": null,
-        "alaskan_native_owned_corporation_or_firm": null,
-        "indian_tribe_federally_recognized": null,
-        "native_hawaiian_owned_business": null,
-        "tribally_owned_business": null,
-        "asian_pacific_american_owned_business": null,
-        "black_american_owned_business": null,
-        "hispanic_american_owned_business": null,
-        "native_american_owned_business": null,
-        "subcontinent_asian_asian_indian_american_owned_business": null,
-        "other_minority_owned_business": null,
-        "us_local_government": null,
-        "undefinitized_action": null,
-        "domestic_or_foreign_entity": null,
-        "domestic_or_foreign_entity_description": null,
-        "division_name": null,
-        "division_number": null,
-        "last_modified_date": "2018-01-19",
-        "certified_date": "2018-01-03",
-        "reporting_period_start": null,
-        "reporting_period_end": null,
-        "create_date": "2018-02-03T16:33:12.744357Z",
-        "update_date": "2018-02-03T16:33:12.744357Z",
-        "city_township_government": null,
-        "special_district_government": null,
-        "small_business": null,
-        "small_business_description": null,
-        "individual": null,
-        "is_fpds": false,
-        "transaction_unique_id": "5_7530_1805NY5MAP_7530-227-1805NY5MAP-5-1-2018-93778-75-0512-NON",
-        "location": {
-            "location_id": 115221974,
-            "data_source": "DBR",
-            "location_country_code": "USA",
-            "country_name": "UNITED STATES",
-            "state_code": "NY",
-            "state_name": "NEW YORK",
-            "state_description": null,
-            "city_name": "ALBANY                      ",
-            "city_code": "01000",
-            "county_name": "ALBANY",
-            "county_code": "001",
-            "address_line1": "NEW YORK STATE DEPARTMENT OF HEALTH  EMPIRE STATE PLAZA",
-            "address_line2": null,
-            "address_line3": null,
-            "foreign_location_description": null,
-            "zip4": null,
-            "zip_4a": null,
-            "congressional_code": "20",
-            "performance_code": null,
-            "zip_last4": "0001",
-            "zip5": "12237",
-            "foreign_postal_code": null,
-            "foreign_province": null,
-            "foreign_city_name": null,
-            "reporting_period_start": null,
-            "reporting_period_end": null,
-            "last_modified_date": "2018-01-19",
-            "certified_date": "2018-01-03",
-            "create_date": "2018-02-03T16:01:24.679252Z",
-            "update_date": "2018-02-03T16:01:24.679252Z",
-            "place_of_performance_flag": false,
-            "recipient_flag": true,
-            "is_fpds": false,
-            "transaction_unique_id": "5_7530_1805NY5MAP_7530-227-1805NY5MAP-5-1-2018-93778-75-0512-NON"
-        },
-        "officers": null
-    },
-    "place_of_performance": {
-        "location_id": 150868557,
-        "data_source": "DBR",
-        "location_country_code": "USA",
-        "country_name": "UNITED STATES",
-        "state_code": "NY",
-        "state_name": "NEW YORK",
-        "state_description": null,
-        "city_name": null,
-        "city_code": null,
-        "county_name": null,
-        "county_code": null,
-        "address_line1": null,
-        "address_line2": null,
-        "address_line3": null,
-        "foreign_location_description": null,
-        "zip4": null,
-        "zip_4a": null,
-        "congressional_code": "90",
-        "performance_code": "NY*****",
-        "zip_last4": null,
-        "zip5": null,
-        "foreign_postal_code": null,
-        "foreign_province": null,
-        "foreign_city_name": null,
-        "reporting_period_start": null,
-        "reporting_period_end": null,
-        "last_modified_date": "2018-01-19",
-        "certified_date": "2018-01-03",
-        "create_date": "2018-02-03T16:01:24.679252Z",
-        "update_date": "2018-02-03T16:01:24.679252Z",
-        "place_of_performance_flag": true,
-        "recipient_flag": false,
-        "is_fpds": false,
-        "transaction_unique_id": "5_7530_1805NY5MAP_7530-227-1805NY5MAP-5-1-2018-93778-75-0512-NON"
-    },
-    "latest_transaction": {
-        "id": 82935535,
-        "usaspending_unique_transaction_id": null,
-        "type": "03",
-        "type_description": null,
-        "period_of_performance_start_date": "2017-10-01",
-        "period_of_performance_current_end_date": "2017-12-31",
-        "action_date": "2018-01-03",
-        "action_type": "A",
-        "action_type_description": null,
-        "federal_action_obligation": "-25000000.00",
-        "original_loan_subsidy_cost": null,
-        "modification_number": "5",
-        "description": "MEDICAID ENTITLEMENT FOR 42 - FY 2018 QUARTER 1 - T19",
-        "drv_award_transaction_usaspend": null,
-        "drv_current_total_award_value_amount_adjustment": null,
-        "drv_potential_total_award_value_amount_adjustment": null,
-        "last_modified_date": "2018-01-19",
-        "certified_date": null,
-        "create_date": "2018-02-03T17:41:52.137156Z",
-        "update_date": "2018-02-03T17:41:52.137156Z",
-        "fiscal_year": 2018,
-        "transaction_unique_id": "5_7530_1805NY5MAP_7530-227-1805NY5MAP-5-1-2018-93778-75-0512-NON",
-        "generated_unique_award_id": "ASST_AW_7530_1805NY5MAP_-NONE-",
-        "is_fpds": false,
-        "award": 44728165,
-        "awarding_agency": {
-            "id": 831,
-            "create_date": "2018-01-31T19:53:04.882146Z",
-            "update_date": "2018-03-07T19:54:56.491927Z",
-            "toptier_flag": false,
-            "toptier_agency": {
-                "toptier_agency_id": 68,
-                "create_date": "2018-01-31T19:53:04.542546Z",
-                "update_date": "2018-03-07T19:54:56.115928Z",
-                "toptier_code": "075",
-                "abbreviation": "HHS",
-                "name": "Department of Health and Human Services",
-                "mission": "It is the mission of the U.S. Department of Health & Human Services (HHS) to enhance and protect the health and well-being of all Americans. We fulfill that mission by providing for effective health and human services and fostering advances in medicine, public health, and social services.",
-                "website": "https://www.hhs.gov/",
-                "icon_filename": "DHHS.jpg"
-            },
-            "subtier_agency": {
-                "subtier_agency_id": 830,
-                "create_date": "2018-01-31T19:53:04.876355Z",
-                "update_date": "2018-03-07T19:54:56.485524Z",
-                "subtier_code": "7530",
-                "abbreviation": "CMS",
-                "name": "Centers for Medicare and Medicaid Services"
-            }
-        },
-        "funding_agency": null,
-        "recipient": {
-            "legal_entity_id": 83053945,
-            "data_source": "DBR",
-            "parent_recipient_unique_id": null,
-            "recipient_name": "NEW YORK STATE DEPARTMENT OF HEALTH",
-            "vendor_doing_as_business_name": null,
-            "vendor_phone_number": null,
-            "vendor_fax_number": null,
-            "business_types": "A",
-            "business_types_description": "STATE GOVERNMENT",
-            "business_categories": [
-                "regional_and_state_government",
-                "government"
-            ],
-            "recipient_unique_id": "806781340",
-            "limited_liability_corporation": null,
-            "sole_proprietorship": null,
-            "partnership_or_limited_liability_partnership": null,
-            "subchapter_scorporation": null,
-            "foundation": null,
-            "for_profit_organization": null,
-            "nonprofit_organization": null,
-            "corporate_entity_tax_exempt": null,
-            "corporate_entity_not_tax_exempt": null,
-            "other_not_for_profit_organization": null,
-            "sam_exception": null,
-            "city_local_government": null,
-            "county_local_government": null,
-            "inter_municipal_local_government": null,
-            "local_government_owned": null,
-            "municipality_local_government": null,
-            "school_district_local_government": null,
-            "township_local_government": null,
-            "us_state_government": null,
-            "us_federal_government": null,
-            "federal_agency": null,
-            "federally_funded_research_and_development_corp": null,
-            "us_tribal_government": null,
-            "foreign_government": null,
-            "community_developed_corporation_owned_firm": null,
-            "labor_surplus_area_firm": null,
-            "small_agricultural_cooperative": null,
-            "international_organization": null,
-            "us_government_entity": null,
-            "emerging_small_business": null,
-            "c8a_program_participant": null,
-            "sba_certified_8a_joint_venture": null,
-            "dot_certified_disadvantage": null,
-            "self_certified_small_disadvantaged_business": null,
-            "historically_underutilized_business_zone": null,
-            "small_disadvantaged_business": null,
-            "the_ability_one_program": null,
-            "historically_black_college": null,
-            "c1862_land_grant_college": null,
-            "c1890_land_grant_college": null,
-            "c1994_land_grant_college": null,
-            "minority_institution": null,
-            "private_university_or_college": null,
-            "school_of_forestry": null,
-            "state_controlled_institution_of_higher_learning": null,
-            "tribal_college": null,
-            "veterinary_college": null,
-            "higher_education": null,
-            "alaskan_native_servicing_institution": null,
-            "community_development_corporation": null,
-            "native_hawaiian_servicing_institution": null,
-            "domestic_shelter": null,
-            "manufacturer_of_goods": null,
-            "hospital_flag": null,
-            "veterinary_hospital": null,
-            "hispanic_servicing_institution": null,
-            "woman_owned_business": null,
-            "minority_owned_business": null,
-            "women_owned_small_business": null,
-            "economically_disadvantaged_women_owned_small_business": null,
-            "joint_venture_women_owned_small_business": null,
-            "joint_venture_economic_disadvantaged_women_owned_small_bus": null,
-            "veteran_owned_business": null,
-            "service_disabled_veteran_owned_business": null,
-            "contracts": null,
-            "grants": null,
-            "receives_contracts_and_grants": null,
-            "airport_authority": null,
-            "council_of_governments": null,
-            "housing_authorities_public_tribal": null,
-            "interstate_entity": null,
-            "planning_commission": null,
-            "port_authority": null,
-            "transit_authority": null,
-            "foreign_owned_and_located": null,
-            "american_indian_owned_business": null,
-            "alaskan_native_owned_corporation_or_firm": null,
-            "indian_tribe_federally_recognized": null,
-            "native_hawaiian_owned_business": null,
-            "tribally_owned_business": null,
-            "asian_pacific_american_owned_business": null,
-            "black_american_owned_business": null,
-            "hispanic_american_owned_business": null,
-            "native_american_owned_business": null,
-            "subcontinent_asian_asian_indian_american_owned_business": null,
-            "other_minority_owned_business": null,
-            "us_local_government": null,
-            "undefinitized_action": null,
-            "domestic_or_foreign_entity": null,
-            "domestic_or_foreign_entity_description": null,
-            "division_name": null,
-            "division_number": null,
-            "last_modified_date": "2018-01-19",
-            "certified_date": "2018-01-03",
-            "reporting_period_start": null,
-            "reporting_period_end": null,
-            "create_date": "2018-02-03T16:33:12.744357Z",
-            "update_date": "2018-02-03T16:33:12.744357Z",
-            "city_township_government": null,
-            "special_district_government": null,
-            "small_business": null,
-            "small_business_description": null,
-            "individual": null,
-            "is_fpds": false,
-            "transaction_unique_id": "5_7530_1805NY5MAP_7530-227-1805NY5MAP-5-1-2018-93778-75-0512-NON",
-            "location": {
-                "location_id": 115221974,
-                "data_source": "DBR",
-                "location_country_code": "USA",
-                "country_name": "UNITED STATES",
-                "state_code": "NY",
-                "state_name": "NEW YORK",
-                "state_description": null,
-                "city_name": "ALBANY                      ",
-                "city_code": "01000",
-                "county_name": "ALBANY",
-                "county_code": "001",
-                "address_line1": "NEW YORK STATE DEPARTMENT OF HEALTH  EMPIRE STATE PLAZA",
-                "address_line2": null,
-                "address_line3": null,
-                "foreign_location_description": null,
-                "zip4": null,
-                "zip_4a": null,
-                "congressional_code": "20",
-                "performance_code": null,
-                "zip_last4": "0001",
-                "zip5": "12237",
-                "foreign_postal_code": null,
-                "foreign_province": null,
-                "foreign_city_name": null,
-                "reporting_period_start": null,
-                "reporting_period_end": null,
-                "last_modified_date": "2018-01-19",
-                "certified_date": "2018-01-03",
-                "create_date": "2018-02-03T16:01:24.679252Z",
-                "update_date": "2018-02-03T16:01:24.679252Z",
-                "place_of_performance_flag": false,
-                "recipient_flag": true,
-                "is_fpds": false,
-                "transaction_unique_id": "5_7530_1805NY5MAP_7530-227-1805NY5MAP-5-1-2018-93778-75-0512-NON"
-            }
-        },
-        "place_of_performance": {
-            "location_id": 150868557,
-            "data_source": "DBR",
-            "location_country_code": "USA",
-            "country_name": "UNITED STATES",
-            "state_code": "NY",
-            "state_name": "NEW YORK",
-            "state_description": null,
-            "city_name": null,
-            "city_code": null,
-            "county_name": null,
-            "county_code": null,
-            "address_line1": null,
-            "address_line2": null,
-            "address_line3": null,
-            "foreign_location_description": null,
-            "zip4": null,
-            "zip_4a": null,
-            "congressional_code": "90",
-            "performance_code": "NY*****",
-            "zip_last4": null,
-            "zip5": null,
-            "foreign_postal_code": null,
-            "foreign_province": null,
-            "foreign_city_name": null,
-            "reporting_period_start": null,
-            "reporting_period_end": null,
-            "last_modified_date": "2018-01-19",
-            "certified_date": "2018-01-03",
-            "create_date": "2018-02-03T16:01:24.679252Z",
-            "update_date": "2018-02-03T16:01:24.679252Z",
-            "place_of_performance_flag": true,
-            "recipient_flag": false,
-            "is_fpds": false,
-            "transaction_unique_id": "5_7530_1805NY5MAP_7530-227-1805NY5MAP-5-1-2018-93778-75-0512-NON"
-        },
-        "assistance_data": {
-            "action_date": "20180103",
-            "action_type": "A",
-            "afa_generated_unique": "5_7530_1805NY5MAP_7530-227-1805NY5MAP-5-1-2018-93778-75-0512-NON",
-            "assistance_type": "03",
-            "award_description": "MEDICAID ENTITLEMENT FOR 42 - FY 2018 QUARTER 1 - T19",
-            "award_modification_amendme": "5",
-            "awardee_or_recipient_legal": "NEW YORK STATE DEPARTMENT OF HEALTH",
-            "awardee_or_recipient_uniqu": "806781340",
-            "awarding_agency_code": "075",
-            "awarding_agency_name": "DEPARTMENT OF HEALTH AND HUMAN SERVICES (HHS)",
-            "awarding_office_code": null,
-            "awarding_office_name": null,
-            "awarding_sub_tier_agency_c": "7530",
-            "awarding_sub_tier_agency_n": "CENTERS FOR MEDICARE AND MEDICAID SERVICES",
-            "business_funds_indicator": "NON",
-            "business_types": "A",
-            "cfda_number": "93.778",
-            "cfda_objectives": "To provide financial assistance to States for payments of medical assistance on behalf of cash assistance recipients, children, pregnant women, and the aged who meet income and resource requirements, and other categorically-eligible groups. In certain States that elect to provide such coverage, medically-needy persons, who, except for income and resources, would be eligible for cash assistance, may be eligible for medical assistance payments under this program. Financial assistance is provided to States to pay for Medicare premiums, copayments and deductibles of qualified Medicare beneficiaries meeting certain income requirements. More limited financial assistance is available for certain Medicare beneficiaries with higher incomes.",
-            "cfda_title": "MEDICAL ASSISTANCE PROGRAM",
-            "correction_delete_indicatr": null,
-            "created_at": "2018-01-19T18:26:40.983841",
-            "face_value_loan_guarantee": null,
-            "fain": "1805NY5MAP",
-            "federal_action_obligation": "-25000000.00",
-            "fiscal_year_and_quarter_co": null,
-            "funding_agency_code": null,
-            "funding_agency_name": null,
-            "funding_office_code": null,
-            "funding_office_name": null,
-            "funding_sub_tier_agency_co": null,
-            "funding_sub_tier_agency_na": null,
-            "is_active": true,
-            "is_historical": null,
-            "legal_entity_address_line1": "NEW YORK STATE DEPARTMENT OF HEALTH  EMPIRE STATE PLAZA",
-            "legal_entity_address_line2": null,
-            "legal_entity_address_line3": null,
-            "legal_entity_city_code": "01000",
-            "legal_entity_city_name": "ALBANY                      ",
-            "legal_entity_congressional": "20",
-            "legal_entity_country_code": "USA",
-            "legal_entity_country_name": "UNITED STATES",
-            "legal_entity_county_code": "001",
-            "legal_entity_county_name": "ALBANY",
-            "legal_entity_foreign_city": null,
-            "legal_entity_foreign_posta": null,
-            "legal_entity_foreign_provi": null,
-            "legal_entity_state_code": "NY",
-            "legal_entity_state_name": "NEW YORK",
-            "legal_entity_zip5": "12237",
-            "legal_entity_zip_last4": "0001",
-            "modified_at": "2018-01-19T18:26:40.982351",
-            "non_federal_funding_amount": "0.00",
-            "original_loan_subsidy_cost": null,
-            "period_of_performance_curr": "20171231",
-            "period_of_performance_star": "20171001",
-            "place_of_perform_country_c": "USA",
-            "place_of_perform_country_n": "UNITED STATES",
-            "place_of_perform_county_co": null,
-            "place_of_perform_county_na": null,
-            "place_of_perform_state_nam": "NEW YORK",
-            "place_of_performance_city": null,
-            "place_of_performance_code": "NY*****",
-            "place_of_performance_congr": "90",
-            "place_of_performance_forei": null,
-            "place_of_performance_zip4a": null,
-            "published_award_financial_assistance_id": "80881516",
-            "record_type": 2,
-            "refresh_from_db": null,
-            "sai_number": "SAI UNAVAILABLE",
-            "total_funding_amount": "-25000000",
-            "transaction": 82935535,
-            "transaction_id": 82935535,
-            "updated_at": "2018-01-19T18:26:40.983848",
-            "uri": "7530-227-1805NY5MAP-5-1-2018-93778-75-0512-NON"
-        },
-        "contract_data": null
-    }
-};
-
-export const mockParams = {
-    awardId: 1234,
-    award: {
-        selectedAward: {
-            category: "contract",
-            id: "ABC1234",
-            internalId: 1234,
-            _startDate: "1999-10-15T04:00:00.000Z",
-            _endDate: "2020-03-31T04:00:00.000Z",
-            _subawardTotal: 0,
-            subawardCount: 0,
-            recipient: {
-                internalId: 5678,
-                name: "Entertainment 720",
-                duns: "222",
-                parentDuns: "333",
-                _businessTypes: "",
-                _businessCategories: ["Minority Owned Business", "Higher Education", "Nonprofit Organization"],
-                location: {
-                    _address1: "602 Trumball Street",
-                    _address2: "",
-                    _address3: "",
-                    _province: "",
-                    _city: "Pawnee",
-                    _county: "Wamapoke",
-                    _stateCode: "IN",
-                    _zip: "00000",
-                    _country: "UNITED STATES",
-                    _countryCode: "",
-                    _state: "IN",
-                    _congressionalDistrict: ""
-                }
-            },
-            placeOfPerformance: {
-                _address1: "",
-                _address2: "",
-                _address3: "",
-                _province: "",
-                _city: "Pawnee",
-                _county: "Wamapoke",
-                _stateCode: "IN",
-                _zip: "00000",
-                _country: "UNITED STATES",
-                _countryCode: "USA",
-                _state: "Indiana",
-                _congressionalDistrict: "03"
-            },
-            awardingAgency: {
-                name: "Department of Sandwiches",
-                subtierName: "Department of Sandwiches",
-                officeName: "Office of Subs"
-            },
-            fundingAgency: {
-                name: "Department of Sandwiches",
-                subtierName: "Department of Sandwiches",
-                officeName: "Office of Subs"
-            },
-            additionalDetails: {
-                pricingCode: "R",
-                pricing: "COST PLUS AWARD FEE",
-                idvType: "--",
-                idcType: "--",
-                idvAgencyId: "--",
-                multipleIdv: "--",
-                solicitationId: "--",
-                solicitationProcedures: "--",
-                numberOffers: "2",
-                extentCompeted: "FULL AND OPEN COMPETITION",
-                notCompeted: "--",
-                setAsideType: "--",
-                commercialAcquisitionProcedures: "COMMERCIAL ITEM PROCEDURES NOT USED",
-                commercialTestProgram: "N",
-                evaluatedPreference: "--",
-                fedBizOpps: "YES",
-                smallBusinessCompetitivenessDemo: "False",
-                _pscCode: "M181",
-                _pscCodeDescription: "OPER OF GOVT R&D GOCO FACILITIES",
-                _naicsCode: "561210",
-                _naicsDescription: "FACILITIES SUPPORT SERVICES",
-                dodClaimantCode: "--",
-                programSystemOrEquipmentCode: "--",
-                itCommercialCategory: "--",
-                seaTransport: "--",
-                clingerCohenAct: "N",
-                constructionWageRateReq: "NOT APPLICABLE",
-                laborStandards: "NOT APPLICABLE",
-                materialSuppliesArticlesEquip: "Y",
-                costOrPricingData: "--",
-                domesticForeign: "--",
-                fairOpportunityLimitedSources: "--",
-                foreignFunding: "NOT APPLICABLE",
-                interagencyContactingAuthority: "ECONOMY ACT",
-                majorProgram: "--",
-                priceEvaluationAdjustmentPreference: "0.00",
-                programAcronym: "--",
-                subcontractingPlan: "PLAN REQUIRED - INCENTIVE NOT INCLUDED ",
-                multiYearContract: "N",
-                purchaseCardAsPaymentMethod: "N",
-                consolidated: "NO",
-                officers: {
-                    officer1: "George Washington - $9,000",
-                    officer2: "John Adams - $7,000",
-                    officer3: "Thomas Jefferson - $6,000",
-                    officer4: "James Madison - $5,000",
-                    officer5: "James Monroe - $5,000"
-                }
-            },
-            parentAward: null,
-            description: "mock description",
-            pricing: "COST PLUS AWARD FEE",
-            _contractType: "DEFINITIVE CONTRACT",
-            _idvType: "--",
-            _amount: 200000.34,
-            _ceiling: 210000.0,
-            _obligation: 204000.9
-        }
-    }
-};
-
-export const mockActions = {
-    setSelectedAward: jest.fn()
-};
