+++ tests/containers/award/mockSearchHelper.js
