--- tests/containers/award/table/mockContractTransaction.js
@@ -1,91 +1,0 @@
-const item = {
-    page_metadata: {
-        page: 1,
-        has_next_page: false
-    },
-    results: [{
-        id: 34779,
-        type: "C",
-        type_description: "Delivery Order",
-        period_of_performance_start_date: "2014-10-01",
-        period_of_performance_current_end_date: "2014-10-30",
-        action_date: "2014-10-01",
-        action_type: null,
-        federal_action_obligation: "227269217.17",
-        modification_number: "0",
-        description: "EXPRESS REPORT PHARMACY PRIME VENDOR VA760PPVFY2015OCT",
-        awarding_agency: {
-            id: 1305,
-            toptier_agency: {
-                toptier_code: "036",
-                abbreviation: "VA",
-                name: "Department of Veterans Affairs"
-            },
-            subtier_agency: {
-                subtier_code: "3600",
-                abbreviation: "",
-                name: "Department of Veterans Affairs"
-            }
-        },
-        funding_agency: {
-            id: 1305,
-            toptier_agency: {
-                toptier_code: "036",
-                abbreviation: "VA",
-                name: "Department of Veterans Affairs"
-            },
-            subtier_agency: {
-                subtier_code: "3600",
-                abbreviation: "",
-                name: "Department of Veterans Affairs"
-            }
-        },
-        recipient: {
-            legal_entity_id: 1576,
-            parent_recipient_unique_id: "177667227",
-            recipient_name: "MCKESSON CORPORATION",
-            business_types: "UN",
-            business_types_description: "Unknown Business Type",
-            location: {
-                country_name: "UNITED STATES",
-                state_code: "CA",
-                state_name: null,
-                city_name: "San Francisco",
-                address_line1: "ONE POST ST",
-                address_line2: "",
-                address_line3: "",
-                zip5: "94104",
-                foreign_postal_code: null,
-                foreign_province: null,
-                foreign_city_name: null,
-                location_country_code: "USA"
-            }
-        },
-        place_of_performance: {
-            country_name: "UNITED STATES",
-            state_code: "CA",
-            state_name: null,
-            city_name: "San Francisco",
-            address_line1: null,
-            address_line2: null,
-            address_line3: null,
-            zip5: "94104",
-            foreign_postal_code: null,
-            foreign_province: null,
-            foreign_city_name: null,
-            location_country_code: "USA"
-        },
-        assistance_data: null,
-        contract_data: {
-            piid: "VA760PPVFY2015OCT",
-            cost_or_pricing_data: "N",
-            type_of_contract_pricing: "J",
-            type_of_contract_pricing_description: "Firm Fixed Price",
-            naics: "NONE",
-            naics_description: null,
-            product_or_service_code: "6505"
-        }
-    }]
-};
-
-export default item;
