--- tests/containers/award/table/mockGrantTransaction.js
@@ -1,79 +1,0 @@
-const items = {
-    page_metadata: {
-        page: 1,
-        has_next_page: false
-    },
-    results: [{
-        id: 40681,
-        type: "05",
-        type_description: "Unknown Type",
-        period_of_performance_start_date: "2007-09-30",
-        period_of_performance_current_end_date: "2017-11-30",
-        action_date: "2016-11-03",
-        action_type: "B",
-        federal_action_obligation: "12393000.00",
-        modification_number: "0022",
-        description: "GREAT LAKES BIOENERGY RESEARCH CENTER",
-        awarding_agency: {
-            id: 1611,
-            toptier_agency: {
-                toptier_code: "089",
-                abbreviation: "DOE",
-                name: "Department of Energy"
-            },
-            subtier_agency: {
-                subtier_code: "8900",
-                abbreviation: "",
-                name: "Department of Energy"
-            }
-        },
-        funding_agency: null,
-        recipient: {
-            legal_entity_id: 2437,
-            parent_recipient_unique_id: "041188822",
-            recipient_name: "UNIVERSITY OF WISCONSIN SYSTEM",
-            business_types: "UN",
-            business_types_description: "Unknown Business Type",
-            location: {
-                country_name: "UNITED STATES",
-                state_code: "WI",
-                state_name: null,
-                city_name: "MADISON",
-                address_line: "21 N PARK ST STE 6401",
-                address_line2: "",
-                address_line3: "",
-                zip5: "53715",
-                foreign_postal_code: null,
-                foreign_province: null,
-                foreign_city_name: null,
-                location_country_code: "USA"
-            }
-        },
-        place_of_performance: {
-            country_name: null,
-            state_code: null,
-            state_name: "WISCONSIN",
-            city_name: "CITY OF MADISON",
-            address_line1: null,
-            address_line2: null,
-            address_line3: null,
-            zip5: null,
-            foreign_postal_code: null,
-            foreign_province: null,
-            foreign_city_name: null,
-            location_country_code: null
-        },
-        assistance_data: {
-            fain: "DEFC0207ER64494",
-            uri: null,
-            cfda_number: "81.049",
-            cfda_title: "OFFICE OF SCIENCE FINANCIAL ASSISTANCE PROGRAM",
-            face_value_loan_guarantee: "0.00",
-            original_loan_subsidy_cost: "0.00",
-            cfda: null
-        },
-        contract_data: null
-    }]
-};
-
-export default items;
