--- tests/containers/award/table/mockLoanTransaction.js
@@ -1,94 +1,0 @@
-const items = {
-    page_metadata: {
-        page: 1,
-        has_next_page: false
-    },
-    results: [{
-        id: 118130,
-        type: "03",
-        type_description: "Formula Grant",
-        period_of_performance_start_date: null,
-        period_of_performance_current_end_date: null,
-        action_date: "2016-12-16",
-        action_type: "B",
-        action_type_description: "Continuation",
-        federal_action_obligation: "220283000.00",
-        modification_number: "0",
-        description: "Medicaid entitlement for Michigan - FY 2017 Quarter 1 - T19",
-        awarding_agency: {
-            id: 802,
-            toptier_agency: {
-                toptier_code: "075",
-                abbreviation: "HHS",
-                name: "Department of Health and Human Services"
-            },
-            subtier_agency: {
-                subtier_code: "7530",
-                abbreviation: "",
-                name: "Centers for Medicare and Medicaid Services"
-            }
-        },
-        funding_agency: {
-            id: 1404,
-            toptier_agency: {
-                toptier_code: "480",
-                abbreviation: "WW1CC",
-                name: "World War I Centennial Commission"
-            },
-            subtier_agency: null
-        },
-        recipient: {
-            legal_entity_id: 113704139,
-            parent_recipient_unique_id: null,
-            recipient_name: "Michigan",
-            business_types: "A",
-            business_types_description: "States Government",
-            location: {
-                country_name: "UNITED STATES",
-                state_code: "MI",
-                state_name: "Michigan",
-                city_name: "LANSING",
-                address_line: "Department of Community Health, Grants Division, 320 S Walnut St",
-                address_line2: null,
-                address_line3: null,
-                zip5: "48933",
-                foreign_postal_code: null,
-                foreign_province: null,
-                foreign_city_name: null,
-                location_country_code: "USA"
-            }
-        },
-        place_of_performance: {
-            country_name: "UNITED STATES",
-            state_code: null,
-            state_name: "Michigan",
-            city_name: null,
-            address_line1: null,
-            address_line2: null,
-            address_line3: null,
-            zip5: null,
-            foreign_postal_code: null,
-            foreign_province: null,
-            foreign_city_name: null,
-            location_country_code: "USA"
-        },
-        assistance_data: {
-            fain: "1705MI5MAP",
-            uri: "7530-227-1705MI5MAP-0-1-2017-93778-75-5512-NON",
-            cfda_number: "93.778",
-            cfda_title: "Medical Assistance Program",
-            face_value_loan_guarantee: null,
-            original_loan_subsidy_cost: null,
-            cfda: {
-                program_number: "93.778",
-                program_title: "Medical Assistance Program",
-                popular_name: "(Medicaid; Title XIX)",
-                objectives: "To provide financial assistance to States for payments of medical assistance on behalf of cash assistance recipients, children, pregnant women, and the aged who meet income and resource requirements, and other categorically-eligible groups. In certain States that elect to provide such coverage, medically-needy persons, who, except for income and resources, would be eligible for cash assistance, may be eligible for medical assistance payments under this program. Financial assistance is provided to States to pay for Medicare premiums, copayments and deductibles of qualified Medicare beneficiaries meeting certain income requirements. More limited financial assistance is available for certain Medicare beneficiaries with higher incomes.",
-                website_address: "http://www.cms.hhs.gov/contracts/."
-            }
-        },
-        contract_data: null
-    }]
-};
-
-export default items;
